### 2024-04-25 #6681 @editorialbot
[PRE REVIEW]: pref_voting: The Preferential Voting Tools package for Python
**Submitting author:** @epacuit (<a href="http://orcid.org/0000-0002-0751-9011">Eric Pacuit</a>)
**Repository:** https://github.com/voting-tools/pref_voting
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
