### 2024-04-27 31f3159
Added Participation axiom
