### 2024-04-28 d0ab60c
add algorithm parameter, add participation axiom
