+++ pref_voting/margin_based_methods_old.py
@@ -15,6 +15,7 @@ import math
 from itertools import product, permutations, combinations, chain
 import networkx as nx
 
+
 @vm(name = "Minimax")
 def minimax(edata, curr_cands = None, strength_function = None):   
     """
@@ -64,7 +65,6 @@ def minimax(edata, curr_cands = None, strength_function = None):
               for c in candidates}
     min_score = min(scores.values())
     return sorted([c for c in candidates if scores[c] == min_score])
-
 
 def minimax_scores(edata, curr_cands = None, score_method="margins"):
     """Return the minimax scores for each candidate, where the minimax score for :math:`c` is -1 * the maximum pairwise majority loss. 
@@ -136,11 +136,13 @@ def maximal_elements(g):
     return [n for n in g.nodes if g.in_degree(n) == 0]
 
 
-def _beat_path_basic(edata, 
-                     curr_cands = None, 
-                     strength_function = None): 
-    """An implementation of the Beat Path method that uses a basic algorithm.  This is not efficient for large graphs.
-    
+@vm(name="Beat Path")
+def beat_path(edata, curr_cands = None, strength_function = None):   
+    """For candidates :math:`a` and :math:`b`, a **path** from :math:`a` to :math:`b` is a sequence 
+    :math:`x_1, \ldots, x_n` of distinct candidates  with  :math:`x_1=a` and :math:`x_n=b` such that 
+    for :math:`1\leq k\leq n-1`, :math:`x_k` is majority preferred to :math:`x_{k+1}`.  The **strength of a path**
+    is the minimal margin along that path.  Say that :math:`a` defeats :math:`b` according to Beat Path if the the strength of the strongest path from :math:`a` to :math:`b` is greater than the strength of the strongest path from :math:`b` to :math:`a`. Then, the candidates that are undefeated according to Beat Path are the winners.  Also known as the Schulze Rule. 
+
     Args:
         edata (Profile, ProfileWithTies, MarginGraph): Any election data that has a `margin` method. 
         curr_cands (List[int], optional): If set, then find the winners for the profile restricted to the candidates in ``curr_cands``
@@ -149,8 +151,37 @@ def _beat_path_basic(edata,
     Returns: 
         A sorted list of candidates. 
 
+    .. seealso::
+
+        :meth:`pref_voting.margin_based_methods.beat_path_Floyd_Warshall`, :meth:`pref_voting.margin_based_methods.beat_path_defeat`
+
+    :Example: 
+
+    .. plot::  margin_graphs_examples/mg_ex_bp_rp.py
+        :context: reset  
+        :include-source: True
+
+
+    .. code-block:: 
+
+        from pref_voting.margin_based_methods import beat_path
+
+        beat_path.display(mg)
+
+
+    .. exec_code:: 
+        :hide_code:
+
+        from pref_voting.weighted_majority_graphs import MarginGraph
+        from pref_voting.margin_based_methods import beat_path
+        
+        mg = MarginGraph([0, 1, 2, 3], [(0, 2, 3), (1, 0, 5), (2, 1, 5), (2, 3, 1), (3, 0, 3), (3, 1, 1)])
+        
+        beat_path.display(mg)
+
+
     """
-    
+
     candidates = edata.candidates if curr_cands is None else curr_cands    
     strength_function = edata.margin if strength_function is None else strength_function
     
@@ -172,10 +203,8 @@ def _beat_path_basic(edata,
     return sorted(list(winners))
 
 
-def _beat_path_floyd_warshall(
-        edata, 
-        curr_cands = None, 
-        strength_function = None):   
+@vm(name="Beat Path")
[…]
