+++ pref_voting/tests/test_margin_based_methods.py
@@ -41,21 +41,7 @@ def condorcet_cycle_prof_with_ties():
         'condorcet_cycle_prof': [0, 1, 2], 
         'condorcet_cycle_prof_with_ties': ['a', 'b', 'c']
     }),
-    (split_cycle_Floyd_Warshall, {
-        'condorcet_cycle': [0, 1, 2], 
-        'cycle': [1], 
-        'linear_margin_graph_0': [0],
-        'linear_margin_graph_0_curr_cands': [1],
-        'condorcet_cycle_prof': [0, 1, 2],         'condorcet_cycle_prof_with_ties': ['a', 'b', 'c']
-    }),
     (beat_path, {
-        'condorcet_cycle': [0, 1, 2], 
-        'cycle': [1], 
-        'linear_margin_graph_0': [0],
-        'linear_margin_graph_0_curr_cands': [1],
-        'condorcet_cycle_prof': [0, 1, 2],         'condorcet_cycle_prof_with_ties': ['a', 'b', 'c']
-    }),
-    (beat_path_Floyd_Warshall, {
         'condorcet_cycle': [0, 1, 2], 
         'cycle': [1], 
         'linear_margin_graph_0': [0],
@@ -70,20 +56,6 @@ def condorcet_cycle_prof_with_ties():
         'condorcet_cycle_prof': [0, 1, 2],         'condorcet_cycle_prof_with_ties': ['a', 'b', 'c']
     }),
     (ranked_pairs_with_test, {
-        'condorcet_cycle': [0, 1, 2], 
-        'cycle': [1], 
-        'linear_margin_graph_0': [0],
-        'linear_margin_graph_0_curr_cands': [1],
-        'condorcet_cycle_prof': [0, 1, 2],         'condorcet_cycle_prof_with_ties': ['a', 'b', 'c']
-    }),
-    # (ranked_pairs_zt, {
-    #     'condorcet_cycle': [0, 1, 2], 
-    #     'cycle': [1], 
-    #     'linear_margin_graph_0': [0],
-    #     'linear_margin_graph_0_curr_cands': [1],
-    #     'condorcet_cycle_prof': [0, 1, 2],         'condorcet_cycle_prof_with_ties': ['a', 'b', 'c']
-    # }),
-    (ranked_pairs_from_stacks, {
         'condorcet_cycle': [0, 1, 2], 
         'cycle': [1], 
         'linear_margin_graph_0': [0],
@@ -125,21 +97,7 @@ def condorcet_cycle_prof_with_ties():
         'linear_margin_graph_0_curr_cands': [1],
         'condorcet_cycle_prof': [0, 1, 2],         'condorcet_cycle_prof_with_ties': ['a', 'b', 'c']
     }),
-    (simple_stable_voting_faster, {
-        'condorcet_cycle': [0, 1, 2], 
-        'cycle': [1], 
-        'linear_margin_graph_0': [0],
-        'linear_margin_graph_0_curr_cands': [1],
-        'condorcet_cycle_prof': [0, 1, 2],         'condorcet_cycle_prof_with_ties': ['a', 'b', 'c']
-    }),
     (stable_voting, {
-        'condorcet_cycle': [0, 1, 2], 
-        'cycle': [1], 
-        'linear_margin_graph_0': [0],
-        'linear_margin_graph_0_curr_cands': [1],
-        'condorcet_cycle_prof': [0, 1, 2],         'condorcet_cycle_prof_with_ties': ['a', 'b', 'c']
-    }),
-    (stable_voting_faster, {
         'condorcet_cycle': [0, 1, 2], 
         'cycle': [1], 
         'linear_margin_graph_0': [0],
@@ -153,13 +111,13 @@ def condorcet_cycle_prof_with_ties():
         'linear_margin_graph_0_curr_cands': [1],
[…]
