### 2024-04-28 76354f5
add missing dois, fix bug in baldwin
