+++ paper/paper.bib
@@ -6,7 +6,8 @@
 	Edition = {2nd},
 	Publisher = {John Wiley \& Sons, Inc.},
 	Title = {Social Choice and Individual Values},
-	Year = {1963}}
+	Year = {1963}, 
+	doi = {https://doi.org/10.12987/9780300186987}}
 	
 	@incollection{Aspremont2002,
 	Address = {Amsterdam},
@@ -53,7 +54,7 @@
 	Publisher = {North-Holland},
 	Title = {Voting Procedures},
 	Volume = {1},
-	doi = {10.1016/s1574-0110(02)80008-x},
+	doi = {https://doi.org/10.1016/S1574-0110(02)80008-X},
 	Year = {2002}}
 	
 		@book{Brams2007,
@@ -62,7 +63,8 @@
 	Publisher = {Springer},
 	Title = {Approval Voting},
 	address = {New York},
-	Year = {2007}}
+	Year = {2007},
+	doi = {https://doi.org/10.1007/978-0-387-49896-6}}
 
 	@book{Balinski2010,
 	Author = {Michel Balinski and Rida Laraki},
@@ -140,7 +142,8 @@
 	Pages = {665-681},
 	Title = {Manipulation of Schemes that Mix Voting with Chance},
 	Volume = {45},
-	Year = {1977}}
+	Year = {1977}, 
+	doi = {https://doi.org/10.2307/1911681}}
 	
 	@article{Green-Armytage2016,
   title={Statistical evaluation of voting rules},
@@ -218,6 +221,7 @@ journal = {Journal of Open Source Software}
 	Publisher = {D. Reidel},
 	Title = {Comparing Voting Systems},
 	Year = {1987},
+	doi = {https://doi.org/10.1007/978-94-009-3985-1}
 	}
 	
 	@book{Nurmi1999,
