+++ pref_voting/tests/test_c1_methods.py
@@ -24,7 +24,9 @@ def condorcet_cycle_prof_with_ties():
 def condorcet_cycle_margin():
     return MarginGraph([0, 1, 2], [(0, 1, 1), (1, 2, 5), (2, 0, 3)])
 
-
+@pytest.fixture
+def profile_single_voter():
+    return Profile([[0, 1, 2]])
 
 @pytest.mark.parametrize("voting_method, expected", [
 (condorcet, {
@@ -33,7 +35,8 @@ def condorcet_cycle_margin():
         'linear_maj_graph_0_curr_cands': [1],
         'condorcet_cycle_prof': [0, 1, 2], 
         'condorcet_cycle_margin': [0, 1, 2],
-        'condorcet_cycle_prof_with_ties': [0, 1, 2]
+        'condorcet_cycle_prof_with_ties': [0, 1, 2],
+        'profile_single_voter': [0]
     }),
     (copeland, {
         'condorcet_cycle': [0, 1, 2], 
@@ -41,7 +44,8 @@ def condorcet_cycle_margin():
         'linear_maj_graph_0_curr_cands': [1],
         'condorcet_cycle_prof': [0, 1, 2], 
         'condorcet_cycle_margin': [0, 1, 2],
-        'condorcet_cycle_prof_with_ties': [0, 1, 2]
+        'condorcet_cycle_prof_with_ties': [0, 1, 2],
+        'profile_single_voter': [0]
     }),
     (uc_gill, {
         'condorcet_cycle': [0, 1, 2], 
@@ -49,7 +53,8 @@ def condorcet_cycle_margin():
         'linear_maj_graph_0_curr_cands': [1],
         'condorcet_cycle_prof': [0, 1, 2], 
         'condorcet_cycle_margin': [0, 1, 2],
-        'condorcet_cycle_prof_with_ties': [0, 1, 2]
+        'condorcet_cycle_prof_with_ties': [0, 1, 2],
+        'profile_single_voter': [0]
     }),
     (llull, {
         'condorcet_cycle': [0, 1, 2], 
@@ -57,7 +62,8 @@ def condorcet_cycle_margin():
         'linear_maj_graph_0_curr_cands': [1],
         'condorcet_cycle_prof': [0, 1, 2], 
         'condorcet_cycle_margin': [0, 1, 2],
-        'condorcet_cycle_prof_with_ties': [0, 1, 2]
+        'condorcet_cycle_prof_with_ties': [0, 1, 2],
+        'profile_single_voter': [0]
     }),
     (top_cycle, {
         'condorcet_cycle': [0, 1, 2], 
@@ -65,7 +71,8 @@ def condorcet_cycle_margin():
         'linear_maj_graph_0_curr_cands': [1],
         'condorcet_cycle_prof': [0, 1, 2], 
         'condorcet_cycle_margin': [0, 1, 2],
-        'condorcet_cycle_prof_with_ties': [0, 1, 2]
+        'condorcet_cycle_prof_with_ties': [0, 1, 2],
+        'profile_single_voter': [0]
     }),
     (uc_fish, {
         'condorcet_cycle': [0, 1, 2], 
@@ -73,7 +80,8 @@ def condorcet_cycle_margin():
         'linear_maj_graph_0_curr_cands': [1],
         'condorcet_cycle_prof': [0, 1, 2], 
         'condorcet_cycle_margin': [0, 1, 2],
-        'condorcet_cycle_prof_with_ties': [0, 1, 2]
+        'condorcet_cycle_prof_with_ties': [0, 1, 2],
+        'profile_single_voter': [0]
     }),
     (uc_bordes, {
         'condorcet_cycle': [0, 1, 2], 
@@ -81,7 +89,8 @@ def condorcet_cycle_margin():
[…]
