+++ pref_voting/variable_voter_axioms.py
@@ -11,6 +11,7 @@ from pref_voting.axiom_helpers import *
 import numpy as np
 from itertools import product, combinations, permutations
 from pref_voting.helper import weak_orders
+from pref_voting.rankings import Ranking
 
 def divide_electorate(prof):
     """Given a Profile or ProfileWithTies object, yield all possible ways to divide the electorate into two nonempty electorates."""
