### 2024-04-28 a60a82e
add DOIs to bib
