+++ paper/paper.bib
@@ -213,8 +213,6 @@ journal = {Journal of Open Source Software}
 	address = {Cambridge, UK}, 
 }
 
-
-
 	@book{Nurmi1987,
 	Address = {Dordrecht},
 	Author = {Hannu Nurmi},
