+++ docs/source/strategic_axioms.md
@@ -11,3 +11,13 @@ Strategic Axioms
 .. autofunction:: pref_voting.strategic_axioms.find_all_strategy_proofness_violations
 
 ```
+
+## Participation
+
+```{eval-rst}
+
+.. autofunction:: pref_voting.variable_voter_axioms.has_participation_violation
+
+.. autofunction:: pref_voting.variable_voter_axioms.find_all_participation_violations
+
+```
