### 2024-04-30 ad9c61e
introduce VotingMethodRegistry, update documentation
