+++ pref_voting/scoring_methods.py
@@ -11,8 +11,17 @@ from pref_voting.profiles import Profile
 from pref_voting.rankings import Ranking, break_ties_alphabetically
 from pref_voting.voting_method import _num_rank_last 
 from pref_voting.profiles import _find_updated_profile, _num_rank
+from pref_voting.voting_method_properties import VotingMethodProperties, ElectionTypes
 
-@vm(name = "Plurality")
+pl_properties = VotingMethodProperties(
+    condorcet_winner=False, 
+    condorcet_loser=False,
+    pareto_dominance=False, 
+    )
+@vm(name = "Plurality", 
+    properties=pl_properties, 
+    input_types=[ElectionTypes.PROFILE, 
+                 ElectionTypes.TRUNCATED_LINEAR_PROFILE])
 def plurality(profile, curr_cands = None):
     """The **Plurality score** of a candidate :math:`c` is the number of voters that rank :math:`c` in first place. The Plurality winners are the candidates with the largest Plurality score in the ``profile`` restricted to ``curr_cands``.
 
@@ -94,7 +103,14 @@ def plurality_ranking(profile, curr_cands=None, local=True, tie_breaking=None):
 
     return p_ranking
 
-@vm(name = "Borda")
+borda_properties = VotingMethodProperties(
+    condorcet_winner=False, 
+    condorcet_loser=True,
+    pareto_dominance=True, 
+    )
+@vm(name = "Borda",
+    properties=borda_properties,
+    input_types=[ElectionTypes.PROFILE])
 def borda(profile, curr_cands = None):
     r"""The **Borda score** of a candidate is calculated as follows: If there are :math:`m` candidates, then the Borda score of candidate :math:`c` is :math:`\sum_{r=1}^{m} (m - r) * Rank(c,r)` where :math:`Rank(c,r)` is the number of voters that rank candidate :math:`c` in position :math:`r`. The Borda winners are the candidates with the largest Borda score in the ``profile`` restricted to ``curr_cands``. 
 
@@ -171,7 +187,14 @@ def borda_ranking(profile, curr_cands=None, local=True, tie_breaking=None):
 
     return b_ranking
 
-@vm(name = "Anti-Plurality")
+anti_plurality_properties = VotingMethodProperties(
+    condorcet_winner=False, 
+    condorcet_loser=False,
+    pareto_dominance=False, 
+    )
+@vm(name = "Anti-Plurality",
+    properties=anti_plurality_properties,
+    input_types=[ElectionTypes.PROFILE])
 def anti_plurality(profile, curr_cands = None):
     """The **Anti-Plurality score** of a candidate $c$ is the number of voters that rank $c$ in last place.  The Anti-Plurality winners are the candidates with the smallest Anti-Plurality score in the ``profile`` restricted to ``curr_cands``. 
 
@@ -247,7 +270,8 @@ def anti_plurality_ranking(profile, curr_cands=None, local=True, tie_breaking=No
     return ap_ranking
 
 
-@vm(name = "Scoring Rule")
+@vm(name = "Scoring Rule",
+    skip_registration=True,)
 def scoring_rule(profile, curr_cands = None, score = lambda num_cands, rank : 1 if rank == 1 else 0):
     """A general scoring rule.  Each voter assign a score to each candidate using the ``score`` function based on their submitted ranking (restricted to candidates in ``curr_cands``).   Returns that candidates with the greatest overall score in the profile restricted to ``curr_cands``. 
 
@@ -309,7 +333,14 @@ def create_scoring_method(score, name):
 
     return VotingMethod(_vm, name = name)
 
-@vm(name = "Dowdall")
+dowdal_properties = VotingMethodProperties(
[…]
