### 2024-05-03 18345c3
Fixed bug in has_smith_violation and find_all_smith_violations
