### 2024-05-03 5c28732
Update other_methods.md
