### 2024-05-03 4ccd4d1
Added marginal algorithm for Borda
