### 2024-05-04 c362dc9
add voting method registry and voting method properties
