+++ pref_voting/combined_methods.py
@@ -16,8 +16,15 @@ from pref_voting.c1_methods import condorcet, smith_set, copeland, top_cycle
 from pref_voting.margin_based_methods import minimax
 from pref_voting.profiles import Profile
 from pref_voting.profiles_with_ties import ProfileWithTies
+from pref_voting.voting_method_properties import VotingMethodProperties, ElectionTypes
 
-@vm(name="Daunou")
+daunou_properties = VotingMethodProperties(
+    condorcet_winner=True,
+    condorcet_loser=True,
+    pareto_dominance=True,)
+@vm(name = "Daunou",
+    properties = daunou_properties,
+    input_types = [ElectionTypes.PROFILE])
 def daunou(profile, curr_cands=None):
     """Implementation of Daunou's voting method as described in the paper: https://link.springer.com/article/10.1007/s00355-020-01276-w
 
@@ -58,7 +65,13 @@ def daunou(profile, curr_cands=None):
     return sorted(winners)
 
 
-@vm(name="Blacks")
+blacks_properties = VotingMethodProperties(
+    condorcet_winner=True,
+    condorcet_loser=True,
+    pareto_dominance=True,)
+@vm(name = "Blacks",
+    properties = blacks_properties,
+    input_types = [ElectionTypes.PROFILE])
 def blacks(profile, curr_cands=None):
     """If a Condorcet winner exists return that winner. Otherwise, return the Borda winning set.
 
@@ -96,7 +109,13 @@ def blacks(profile, curr_cands=None):
 
     return winners
 
-@vm(name="Smith IRV")
+smith_irv_properties = VotingMethodProperties(
+    condorcet_winner=True,
+    condorcet_loser=True,
+    pareto_dominance=True,)
+@vm(name = "Smith IRV",
+    properties = smith_irv_properties,
+    input_types = [ElectionTypes.PROFILE])
 def smith_irv(profile, curr_cands=None):
     """After restricting to the Smith Set, return the Instant Runoff winner.
 
@@ -129,7 +148,13 @@ def smith_irv(profile, curr_cands=None):
 
     return instant_runoff(profile, curr_cands=smith)
 
-@vm(name="Smith IRV PUT")
+smith_irv_put_properties = VotingMethodProperties(
+    condorcet_winner=True,
+    condorcet_loser=True,
+    pareto_dominance=True,)
+@vm(name = "Smith IRV PUT",
+    properties = smith_irv_put_properties,
+    input_types = [ElectionTypes.PROFILE])
 def smith_irv_put(profile, curr_cands=None):
     """After restricting to the Smith Set, return the Instant Runoff winner.
 
@@ -162,8 +187,13 @@ def smith_irv_put(profile, curr_cands=None):
 
     return instant_runoff_put(profile, curr_cands=smith)
 
-
-@vm(name="Condorcet IRV")
+condorcet_irv_properties = VotingMethodProperties(
+    condorcet_winner=True,
+    condorcet_loser=True,
+    pareto_dominance=True,)
+@vm(name = "Condorcet IRV",
+    properties = condorcet_irv_properties,
+    input_types = [ElectionTypes.PROFILE])
 def condorcet_irv(profile, curr_cands=None):
     """If a Condorcet winner exists, elect that candidate, otherwise return the instant runoff winners.
 
@@ -198,7 +228,13 @@ def condorcet_irv(profile, curr_cands=None):
     else:
         return instant_runoff(profile, curr_cands=curr_cands)
 
-@vm(name="Condorcet IRV PUT")
+condorcet_irv_put_properties = VotingMethodProperties(
+    condorcet_winner=True,
+    condorcet_loser=True,
+    pareto_dominance=True,)
+@vm(name = "Condorcet IRV PUT",
+    properties = condorcet_irv_put_properties,
[…]
