### 2024-05-05 d6f5d5a
Added Dowdall to scoring method methods
