### 2024-05-05 48f0cb9
Added positive_involvement to voting method properties
