+++ pref_voting/iterative_methods.py
@@ -71,6 +71,7 @@ irv_properties = VotingMethodProperties(
     condorcet_winner=False, 
     condorcet_loser=False,
     pareto_dominance=True, 
+    positive_involvement=True,
     )
 @vm(name = "Instant Runoff",
     properties=irv_properties,
@@ -244,6 +245,7 @@ irv_put_properties = VotingMethodProperties(
     condorcet_winner=False, 
     condorcet_loser=None,
     pareto_dominance=True, 
+    positive_involvement=True,
     )
 @vm(name = "Instant Runoff PUT",
     properties=irv_put_properties,
@@ -512,6 +514,7 @@ bottom_two_runoff_properties = VotingMethodProperties(
     condorcet_winner=True, 
     condorcet_loser=False,
     pareto_dominance=True, 
+    positive_involvement=False,
     )
 @vm(name="Bottom-Two-Runoff Instant Runoff",
     properties=bottom_two_runoff_properties,
@@ -569,6 +572,7 @@ bottom_two_runoff_put_properties = VotingMethodProperties(
     condorcet_winner=None, 
     condorcet_loser=False,
     pareto_dominance=None, 
+    positive_involvement=False,
     )
 @vm(name="Bottom-Two-Runoff Instant Runoff PUT",
     properties=bottom_two_runoff_put_properties,
@@ -620,6 +624,7 @@ pl_w_runoff_properties = VotingMethodProperties(
     condorcet_winner=False, 
     condorcet_loser=True,
     pareto_dominance=None, 
+    positive_involvement=True,
     )
 @vm(name = "PluralityWRunoff PUT",
     properties=pl_w_runoff_properties,
@@ -734,6 +739,7 @@ coombs_properties = VotingMethodProperties(
     condorcet_winner=False, 
     condorcet_loser=True,
     pareto_dominance=True, 
+    positive_involvement=False,
     )
 @vm(name = "Coombs",
     properties=coombs_properties,
@@ -806,6 +812,7 @@ coombs_tb_properties = VotingMethodProperties(
     condorcet_winner=False, 
     condorcet_loser=None,
     pareto_dominance=True, 
+    positive_involvement=False,
     )
 @vm(name = "Coombs TB",
     properties=coombs_tb_properties,
@@ -883,6 +890,7 @@ coombs_put_properties = VotingMethodProperties(
     condorcet_winner=False, 
     condorcet_loser=None,
     pareto_dominance=True, 
+    positive_involvement=False,
     )
 @vm(name = "Coombs PUT",
     properties=coombs_put_properties,
@@ -1022,6 +1030,7 @@ baldwin_properties = VotingMethodProperties(
     condorcet_winner=True, 
     condorcet_loser=True,
     pareto_dominance=True, 
+    positive_involvement=False,
     )
 @vm(name = "Baldwin",
     properties=baldwin_properties,
@@ -1096,6 +1105,7 @@ baldwin_tb_properties = VotingMethodProperties(
     condorcet_winner=True, 
     condorcet_loser=True,
     pareto_dominance=True, 
+    positive_involvement=False,
     )
 @vm(name = "Baldwin TB",
     properties=coombs_tb_properties,
@@ -1186,6 +1196,7 @@ baldwin_put_properties = VotingMethodProperties(
     condorcet_winner=True, 
     condorcet_loser=True,
     pareto_dominance=True, 
+    positive_involvement=False,
     )
 @vm(name = "Baldwin PUT",
     properties=baldwin_put_properties,
@@ -1333,6 +1344,7 @@ strict_nanson_properties = VotingMethodProperties(
     condorcet_winner=True, 
     condorcet_loser=True,
     pareto_dominance=True, 
+    positive_involvement=False,
     )
[…]
