### 2024-05-05 160f0cc
Update margin_based_methods.py
