### 2024-05-05 134f1d0
Added schwartz_sequential_dropping algorithm for beat_path
