### 2024-05-05 73fb061
Improved speed of generate_edge_ordered_tournament
