+++ pref_voting/generate_weighted_majority_graphs.py
@@ -29,17 +29,14 @@ def generate_edge_ordered_tournament(num_cands, parity="even"):
     assert parity in ["even", "odd"], "The parity should be either 'even' or 'odd'."
 
     mg = nx.DiGraph()
-    mg.add_nodes_from(list(range(num_cands)))
-    candidates = list(range(num_cands))
+    mg.add_nodes_from(range(num_cands))
     _edges = list()
-    for c1 in candidates: 
-        for c2 in candidates: 
-            if c1 != c2: 
-                if (c1, c2) not in _edges and (c2, c1) not in _edges:
-                    if np.random.choice([True, False]): 
-                        _edges.append((c1, c2))
-                    else: 
-                        _edges.append((c2, c1))
+    for c1 in range(num_cands): 
+        for c2 in range(c1+1, num_cands): 
+            if np.random.choice([True, False]): 
+                _edges.append((c1, c2))
+            else: 
+                _edges.append((c2, c1))
                    
     edges = list()
     edge_indices = list(range(len(_edges)))
@@ -48,7 +45,7 @@ def generate_edge_ordered_tournament(num_cands, parity="even"):
     for i, e_idx in enumerate(edge_indices):
         edges.append((_edges[e_idx][0], _edges[e_idx][1], 2 * (i+1) if parity == 'even' else 2 * i+1)) 
     
-    return MarginGraph(candidates, edges)
+    return MarginGraph(range(num_cands), edges)
 
 def generate_margin_graph(num_cands, weight_domain = None, parity = 'even'): 
     """Generate a random MarginGraph (allowing for ties in the margins) for ``num_cands`` candidates.  
