+++ docs/source/conf.py
@@ -22,7 +22,7 @@ copyright = '2023, Wes Holliday and Eric Pacuit'
 author = 'Wes Holliday and Eric Pacuit'
 
 # The full version, including alpha/beta/rc tags
-release = '1.3.0'
+release = '1.3.2'
 
 
 # -- General configuration ---------------------------------------------------
