### 2024-05-23 e418e19
add voting method properties
