+++ pref_voting/other_methods.py
@@ -13,7 +13,7 @@ from pref_voting.profiles import _find_updated_profile, _num_rank
 from pref_voting.profiles_with_ties import ProfileWithTies
 from pref_voting.weighted_majority_graphs import MarginGraph
 from itertools import combinations, permutations
-from pref_voting.voting_method_properties import VotingMethodProperties, ElectionTypes
+from pref_voting.voting_method_properties import ElectionTypes
 
 @vm(name = "Majority",
     skip_registration=True, # skip registration since majority may return an empty list
@@ -54,14 +54,7 @@ def majority(profile, curr_cands = None):
 
     return sorted(maj_winner)
 
-pareto_properties = VotingMethodProperties(
-    condorcet_winner=False,
-    condorcet_loser=False,
-    pareto_dominance=True,
-    positive_involvement = True,
-)
 @vm(name = "Pareto",
-    properties = pareto_properties,
     input_types = [ElectionTypes.PROFILE, ElectionTypes.PROFILE_WITH_TIES])
 def pareto(profile, curr_cands = None, strong_Pareto = False, use_extended_strict_preferences = True):
     """Returns the set of candidates who are not Pareto dominated.
@@ -164,14 +157,7 @@ def kemeny_young_rankings(profile, curr_cands = None):
     return _kemeny_young_rankings(list(rankings), list(profile._rcounts), candidates)
 
 
-kemeny_young_properties = VotingMethodProperties(
-    condorcet_winner=True,
-    condorcet_loser=True,
-    pareto_dominance=True,
-    positive_involvement=False,
-)
 @vm(name = "Kemeny-Young",
-    properties = kemeny_young_properties,
     input_types = [ElectionTypes.PROFILE,ElectionTypes.PROFILE_WITH_TIES,  ElectionTypes.MARGIN_GRAPH])
 def kemeny_young(edata, curr_cands = None, algorithm = "marginal"): 
     """A Kemeny-Young ranking is a ranking that maximizes the sum of the margins of pairs of candidates in the ranking. Equivalently, a Kemeny-Young ranking is a ranking that minimizes the sum of the Kendall tau distances to the voters' rankings. The Kemeny-Young winners are the candidates that are ranked first by some Kemeny-Young ranking.
@@ -239,14 +225,7 @@ def kemeny_young(edata, curr_cands = None, algorithm = "marginal"):
     
     return sorted(list(set([r[0] for r in ky_rankings])))
 
-preliminary_weighted_condorcet_properties = VotingMethodProperties(
-    condorcet_winner=True,
-    condorcet_loser=True,
-    pareto_dominance=True,
-    positive_involvement=False,
-)
 @vm("Preliminary Weighted Condorcet",
-    properties = preliminary_weighted_condorcet_properties,
     input_types = [ElectionTypes.PROFILE])
 def preliminary_weighted_condorcet(prof, curr_cands = None, show_orders = False, require_positive_plurality_score = False):
     """The preliminary version of the Weighted Condorcet Rule in Tideman's book, Collective Decisions and Voting (p. 223). The winners are the candidates ranked first by some linear order of the candidates with highest score, where the score of an order (c_1,...,c_n) is the sum over all i<j of the margin of c_i vs. c_j multiplied by the plurality scores of c_i and c_j. 
@@ -295,14 +274,7 @@ def preliminary_weighted_condorcet(prof, curr_cands = None, show_orders = False,
 
 ### Bucklin
 
-bucklin_properties = VotingMethodProperties(
-    condorcet_winner=False,
-    condorcet_loser=False,
-    pareto_dominance=True,
-    positive_involvement=False,
-)
 @vm(name = "Bucklin",
-    properties = bucklin_properties,
     input_types = [ElectionTypes.PROFILE])
 def bucklin(profile, curr_cands = None): 
[…]
