+++ pref_voting/c1_methods.py
@@ -18,16 +18,9 @@ import math
 from itertools import product, permutations, combinations, chain
 import networkx as nx
 import matplotlib.pyplot as plt
-from pref_voting.voting_method_properties import VotingMethodProperties, ElectionTypes
+from pref_voting.voting_method_properties import ElectionTypes
 
-condorcet_properties = VotingMethodProperties(
-    condorcet_winner=True,
-    condorcet_loser=False,
-    pareto_dominance=False,
-    positive_involvement=False,
-    )
 @vm(name = "Condorcet",
-    properties = condorcet_properties,
     input_types = [ElectionTypes.PROFILE, ElectionTypes.PROFILE_WITH_TIES, ElectionTypes.MAJORITY_GRAPH, ElectionTypes.MARGIN_GRAPH])
 def condorcet(edata, curr_cands = None):
     """
@@ -74,14 +67,7 @@ def condorcet(edata, curr_cands = None):
     
     return [cond_winner] if cond_winner is not None else sorted(candidates)
 
-copeland_properties = VotingMethodProperties(
-    condorcet_winner=True,
-    condorcet_loser=True,
-    pareto_dominance=True,
-    positive_involvement=False,
-    )
 @vm(name = "Copeland",
-    properties = copeland_properties,
     input_types = [ElectionTypes.PROFILE, ElectionTypes.PROFILE_WITH_TIES, ElectionTypes.MAJORITY_GRAPH, ElectionTypes.MARGIN_GRAPH])
 def copeland(edata, curr_cands = None):
     """The Copeland score for c is the number of candidates that c is majority preferred to minus the number of candidates majority preferred to c.  The Copeland winners are the candidates with the maximum Copeland score in the profile restricted to ``curr_cands``. 
@@ -163,14 +149,7 @@ def copeland_ranking(edata, curr_cands=None, local=True, tie_breaking=None):
     return copeland_ranking
 
 
-llull_properties = VotingMethodProperties(
-    condorcet_winner=True,
-    condorcet_loser=True,
-    pareto_dominance=True,
-    positive_involvement=False,
-    )
 @vm(name = "Llull",
-    properties = llull_properties,
     input_types = [ElectionTypes.PROFILE, ElectionTypes.PROFILE_WITH_TIES, ElectionTypes.MAJORITY_GRAPH, ElectionTypes.MARGIN_GRAPH])
 def llull(edata, curr_cands = None):
     """The Llull score for a candidate :math:`c` is the number of candidates that :math:`c` is weakly majority preferred to.  This is equivalent to calculating the Copeland scores for a candidate :math:`c` with 1 point for each candidate that :math:`c` is majority preferred to, 1/2 point for each candidate that :math:`c` is tied with, and 0 points for each candidate that is majority preferred to :math:`c`.  The Llull winners are the candidates with the maximum Llull score in the profile restricted to ``curr_cands``. 
@@ -230,13 +209,7 @@ def right_covers(dom, c1, c2):
       
     return dom[c2].issubset(dom[c1])
 
-uc_gill_properties = VotingMethodProperties(
-    condorcet_winner=True,
-    condorcet_loser=True,
-    pareto_dominance=True,)
-
 @vm(name = "Uncovered Set",
-    properties = uc_gill_properties,
     input_types = [ElectionTypes.PROFILE, ElectionTypes.PROFILE_WITH_TIES, ElectionTypes.MAJORITY_GRAPH, ElectionTypes.MARGIN_GRAPH])
 def uc_gill(edata, curr_cands = None): 
[…]
