### 2024-05-25 ad53d03
various updates to improve interaction with voting_methods
