### 2024-05-29 74ba65c
Added Bradley-Terry as a vm and swf
