+++ pref_voting/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '1.3.8'
+__version__ = '1.3.11'
