### 2024-05-31 7b727d3
Added references for stability for winners and immunity to spoilers.
