### 2024-05-31 c957535
Changes majority vm to absolute_majority
