### 2024-05-31 2003ab2
update some references to spoiler axioms, add condorcet violation instances, change majority to absolute_majority
