+++ docs/source/other_methods.md
@@ -1,7 +1,7 @@
 Other Methods
 =======================================
 
-## Majority
+## Absolute Majority
 
 ```{eval-rst}
 
