### 2024-06-11 34f62fe
Added Copeland-Global-Minimax
