### 2024-06-28 #6681 @britta-wstnr
PS: @epacuit our trusty editorialbot has flagged a few DOIs as missing in your paper, see [here](https://github.com/openjournals/joss-reviews/issues/6681#issuecomment-2077157686). Can you have a look?
