### 2024-06-30 a7de75b
version 1.3.20, add DOIs to the bib for the paper.
