+++ paper/paper.bib
@@ -7,7 +7,7 @@
 	Publisher = {John Wiley \& Sons, Inc.},
 	Title = {Social Choice and Individual Values},
 	Year = {1963}, 
-	doi = {https://doi.org/10.12987/9780300186987}}
+	doi = {10.12987/9780300186987}}
 	
 	@incollection{Aspremont2002,
 	Address = {Amsterdam},
@@ -39,6 +39,7 @@
   title={Guide to Numerical Experiments on Elections in Computational Social Choice},
   author={Boehmer, N. and Faliszewski, P. and Janeczko, \L. and Kaczmarczyk, A. and Lisowski, G. and Pierczy\'{n}ski, G. and Rey, S. and Stolicki, D. and Szufa, S. and W\k{a}s, T.},
   year={2024},
+  url={https://arxiv.org/abs/2402.11765},
   eprint={2402.11765},
   archivePrefix={arXiv},
   primaryClass={cs.GT}
@@ -54,7 +55,7 @@
 	Publisher = {North-Holland},
 	Title = {Voting Procedures},
 	Volume = {1},
-	doi = {https://doi.org/10.1016/S1574-0110(02)80008-X},
+	doi = {10.1016/S1574-0110(02)80008-X},
 	Year = {2002}}
 	
 		@book{Brams2007,
@@ -71,13 +72,15 @@
 	Publisher = {MIT Press},
 	Address = {Boston},
 	Title = {Majority Judgement: Measuring, Ranking and Electing},
-	Year = {2010}}
+	Year = {2010},
+	doi={10.7551/mitpress/9780262015134.003.0001}}
 	
 	@book{Brandt2016,
 	title = {Handbook of Computational Social Choice},
 	Publisher = {Cambridge University Press},
 	Editor = {Felix Brandt and Vincent Conitzer and Ulle Endriss and J\'{e}r\^{o}me Lang and Ariel D. Procaccia},
-	Year = 2016}	
+	Year = 2016,
+	doi={10.1017/CBO9781107446984}}	
 
 	
 		@book{Dummett1984,
@@ -115,7 +118,8 @@
 	Publisher = {Princeton University Press},
 	address = {Princeton, New Jersey},
 	Title = {The Theory of Social Choice},
-	Year = {1973}}
+	Year = {1973},
+	doi = {10.1515/9781400868339}}
 	
 	@inproceedings{Geist2017,
 	Author = {Christian Geist and Dominik Peters},
@@ -124,7 +128,8 @@
 	Pages = {249-267},
 	Publisher = {AI Access},
 	Title = {Computer-aided Methods for Social Choice Theory},
-	Year = {2017}}
+	Year = {2017},
+	url={https://archive.illc.uva.nl/COST-IC1205/BookDocs/Chapters/TrendsCOMSOC-13.pdf}}
 	
   @inproceedings{Brandt2017,
 	Author = {Felix Brandt},
@@ -133,7 +138,8 @@
 	Pages = {3-26},
   Publisher = {AI Access},
 	Title = {Rolling the dice: Recent results in probabilistic social choice},
-	Year = {2017}
+	Year = {2017},
+	url={https://archive.illc.uva.nl/COST-IC1205/BookDocs/Chapters/TrendsCOMSOC-01.pdf}
 }
 
 	@article{Gibbard1977,
@@ -143,7 +149,7 @@
 	Title = {Manipulation of Schemes that Mix Voting with Chance},
 	Volume = {45},
 	Year = {1977}, 
-	doi = {https://doi.org/10.2307/1911681}}
+	doi = {10.2307/1911681}}
 	
 	@article{Green-Armytage2016,
   title={Statistical evaluation of voting rules},
[…]
