+++ pref_voting/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '1.13.17'
+__version__ = '1.13.20'
