### 2024-06-30 #6681 @epacuit
@britta-wstnr: We have updated all of the DOIs in the article:

- 10.12987/9780300186987 may be a valid DOI for title: Social Choice and Individual Values

Fixed

- No DOI given, and none found for title: Social welfare functionals and interpersonal compa...

Added a DOI

- No DOI given, and none found for title: Guide to Numerical Experiments on Elections in Com...

No DOI, but added link to the article on arXiv

- 10.7551/mitpress/9780262015134.003.0018 may be a valid DOI for title: Approval Voting

Fixe

- No DOI given, and none found for title: Majority Judgement: Measuring, Ranking and Electin...

Added a DOI

- No DOI given, and none found for title: Handbook of Computational Social Choice

Added a DOI

- 10.1093/oso/9780190934163.003.0003 may be a valid DOI for title: Voting Procedures

This is not the correct DOI for our citation.  There is no DOI for this book.

- No DOI given, and none found for title: The Theory of Social Choice

Added a DOI

- No DOI given, and none found for title: Computer-aided Methods for Social Choice Theory

Added a link to the paper (there is no DOI for this article).

- No DOI given, and none found for title: Rolling the dice: Recent results in probabilistic ...

Added a link to the paper (there is no DOI for this article).

- 10.2307/1911681 may be a valid DOI for title: Manipulation of Schemes that Mix Voting with Chanc...

Fixed

- No DOI given, and none found for title: Learning to Manipulate under Limited Information

Added a link to the paper on arXiv (there is no DOI for this article).

- No DOI given, and none found for title: Making Multicandidate Elections More Democratic

DOI added

- No DOI given, and none found for title: A Unified Theory of Voting: Directional and Proxim...

DOI added

- 10.1007/978-94-009-3985-1 may be a valid DOI for title: Comparing Voting Systems

Fixed

- No DOI given, and none found for title: Voting Methods

There is no DOI for this article, but a link to the online article is provided.

- No DOI given, and none found for title: Collective Choice and Social Welfare: An Expanded ...

DOI added.

- No DOI given, and none found for title: Votelib: Evaluation of voting systems in Python

There is no DOI for this, but a link to the GitHub repository is provide.

- No DOI given, and none found for title: Pref.Tools: Tools that are useful for analyzing pr...

There is no DOI for this, but a link to the website is provided.

- No DOI given, and none found for title: abif

There is no DOI for this, but a link to the GitHub repository is provided.

- No DOI given, and none found for title: VoteKit

There is no DOI for this, but a link to the GitHub repository is provided.
