### 2024-06-30 e8d41a9
format the bib file
