+++ paper/paper.bib
@@ -1,325 +1,327 @@
-	
-	
-	@book{Arrow1963,
-	Address = {New York},
-	Author = {Kenneth J. Arrow},
-	Edition = {2nd},
-	Publisher = {John Wiley \& Sons, Inc.},
-	Title = {Social Choice and Individual Values},
-	Year = {1963}, 
-	doi = {10.12987/9780300186987}}
-	
-	@incollection{Aspremont2002,
-	Address = {Amsterdam},
-	Author = {Claude d'Aspremont and Louis Gevers},
-	Booktitle = {Handbook of Social Choice and Welfare},
-	Editor = {Kenneth J. Arrow and Amartya K. Sen and Kotaro Suzumura},
-	Pages = {459-541},
-	Publisher = {North-Holland},
-	Title = {Social welfare functionals and interpersonal comparability},
-	Volume = {1},
-	Year = {2002},
-	URL = {https://doi.org/10.1016/S1574-0110(02)80014-5},
-	}
-	
-	@InCollection{Aziz2019,
-  author    = {Haris Aziz and Felix Brandt and Edith Elkind and Piotr Skowron},
-  title     = {Computational Social Choice: The First Ten Years and Beyond},
-  booktitle = {Computing and Software Science},
-  publisher = {Springer},
-  year      = {2019},
-  editor    = {Bernhard Steffen and Gerhard Woeginger},
-  volume    = {10000},
-  series    = {Lecture Notes in Computer Science},
-  doi       = {10.1007/978-3-319-91908-9_4},
-}
-	
-	
-	@misc{Boehmer2024,
-  title={Guide to Numerical Experiments on Elections in Computational Social Choice},
-  author={Boehmer, N. and Faliszewski, P. and Janeczko, \L. and Kaczmarczyk, A. and Lisowski, G. and Pierczy\'{n}ski, G. and Rey, S. and Stolicki, D. and Szufa, S. and W\k{a}s, T.},
-  year={2024},
-  url={https://arxiv.org/abs/2402.11765},
-  eprint={2402.11765},
-  archivePrefix={arXiv},
-  primaryClass={cs.GT}
+@book{Arrow1963,
+  title      = {Social Choice and Individual Values},
+  author     = {Kenneth J. Arrow},
+  address    = {New York},
+  edition    = {2nd},
+  publisher  = {John Wiley \& Sons, Inc.},
+  year       = {1963},
+  doi        = {10.12987/9780300186987}
 }
 
-
-	@incollection{Brams2002,
-	Address = {Amsterdam},
-	Author = {Steven J. Brams and Peter C. Fishburn},
-	Booktitle = {Handbook of Social Choice and Welfare},
-	Editor = {Kenneth J. Arrow and Amartya K. Sen and Kotaro Suzumura},
-	Pages = {173-236},
-	Publisher = {North-Holland},
-	Title = {Voting Procedures},
-	Volume = {1},
-	doi = {10.1016/S1574-0110(02)80008-X},
-	Year = {2002}}
-	
-		@book{Brams2007,
-	Author = {Steven J. Brams and Peter C. Fishburn},
-	Edition = {2nd},
-	Publisher = {Springer},
-	Title = {Approval Voting},
-	address = {New York},
-	Year = {2007},
-	doi = {https://doi.org/10.1007/978-0-387-49896-6}}
-
-	@book{Balinski2010,
-	Author = {Michel Balinski and Rida Laraki},
-	Publisher = {MIT Press},
-	Address = {Boston},
-	Title = {Majority Judgement: Measuring, Ranking and Electing},
-	Year = {2010},
-	doi={10.7551/mitpress/9780262015134.003.0001}}
-	
-	@book{Brandt2016,
-	title = {Handbook of Computational Social Choice},
-	Publisher = {Cambridge University Press},
[…]
