### 2024-07-02 b25efb7
fix doi in bib file for JOSS paper
