+++ paper/paper.bib
@@ -63,7 +63,7 @@
   publisher  = {Springer},
   address    = {New York},
   year       = {2007},
-  doi        = {https://doi.org/10.1007/978-0-387-49896-6}
+  doi        = {10.1007/978-0-387-49896-6}
 }
 
 @book{Balinski2010,
