### 2024-07-02 #6681 @epacuit
Here are two people that seem like they might be good reviewers:

Stephen Francis Mann: https://reviewers.joss.theoj.org/reviewers/3439
Maximilian Stahlberg: https://reviewers.joss.theoj.org/reviewers/1964

Another suggestion from the computational social choice community is Martin Lackner (https://martin.lackner.xyz/).   He is the main developer for the abcvoting package (https://abcvoting.readthedocs.io/en/latest/index.html), which is a related Python package, with a JOSS paper: https://joss.theoj.org/papers/10.21105/joss.04880

Some additional suggestions are the reviewers for the abcvoting JOSS paper:

Stefan McCabe: https://github.com/sdmccabe
Christopher Barrie: https://github.com/cjbarrie

Another suggestion from the computational social choice community is Dominik Peters (https://dominik-peters.de/).
