### 2024-07-08 #6681 @Viech
Maximilian Stahlberg here. I'm always happy to review for JOSS but my defense is coming up, so I will only be available from the 25th onward. If you are still short by then, don't hesitate to give me another ping!
