### 2024-07-16 7e94a82
add to_utility_profile with uncertainty about candidate location; add tops_scores to ProfileWithTies
