+++ pref_voting/spatial_profiles.py
@@ -3,7 +3,6 @@ import matplotlib.pyplot as plt
 import seaborn as sns
 from pref_voting.utility_functions import *
 from pref_voting.utility_profiles import UtilityProfile
-
 class SpatialProfile(object): 
     """
     A spatial profile is a set of candidates and voters in a multi-dimensional space.  Each voter and candidate is assigned vector of floats representing their position on each issue.
@@ -49,23 +48,54 @@ class SpatialProfile(object):
         """
         return self.cand_pos[c]
     
-    def to_utility_profile(self, utility_function = None):
+    def to_utility_profile(self, 
+                           utility_function = None,
+                           uncertainty_function=None,
+                           batch=False, 
+                           return_virtual_cand_positions=False):
         """
         Returns a utility profile corresponding to the spatial profile.  
         
         Args:
-            utility_function (optional, function): A function that takes two vectors and returns a float. The default utility function is the quadratic utility function.
-
+            utility_function (callable, optional): A function that takes two vectors and returns a float. The default utility function is the quadratic utility function.
+            uncertainty_function (callable, optional): A function that models uncertainty and returns covariance parameters.
+            batch (bool, optional): If True, generate positions in batches. Default is False.
+            return_virtual_cand_positions (bool, optional): If True, return virtual candidate positions. Default is False.
+            
         Returns:    
             UtilityProfile: A utility profile corresponding to the spatial profile.
-
+            (optional) Tuple[UtilityProfile, dict]: The utility profile and virtual candidate positions if `return_virtual_cand_positions` is True.
         """
-        utility_function = quadratic_utility if utility_function is None else utility_function 
+        import numpy as np
+        from pref_voting.generate_spatial_profiles import generate_covariance
 
-        return UtilityProfile([
-            {c:utility_function(self.voter_position(v), self.candidate_position(c)) 
-             for c in self.candidates} for v in self.voters
-            ])
+        utility_function = utility_function or quadratic_utility
+
+        if uncertainty_function is not None:
+            virtual_cand_positions = {}
+            for c in self.candidates:
+                if batch:
+                    covariance = generate_covariance(self.num_dims, *uncertainty_function(self, c, self.voters[0]))
+                    positions = np.random.multivariate_normal(self.candidate_position(c), covariance, size=len(self.voters))
+                else:
+                    positions = [np.random.multivariate_normal(self.candidate_position(c), generate_covariance(self.num_dims, *uncertainty_function(self, c, v))) for v in self.voters]
+                virtual_cand_positions[c] = positions
+            
+            utility_profile = [
+                {c: utility_function(self.voter_position(v), virtual_cand_positions[c][vidx]) for c in self.candidates}
+                for vidx, v in enumerate(self.voters)
+            ]
+            
+            if return_virtual_cand_positions:
+                return UtilityProfile(utility_profile), virtual_cand_positions
+            else:
+                return UtilityProfile(utility_profile)
+        else:
+            utility_profile = [
+                {c: utility_function(np.array(self.voter_position(v)), np.array(self.candidate_position(c))) for c in self.candidates}
+                for v in self.voters
+            ]
+            return UtilityProfile(utility_profile)
     
     def to_string(self): 
         """
