### 2024-07-22 #7020 @editorialbot
[REVIEW]: pref_voting: The Preferential Voting Tools package for Python
**Submitting author:** @epacuit (<a href="http://orcid.org/0000-0002-0751-9011">Eric Pacuit</a>)
**Repository:** https://github.com/voting-tools/pref_voting
**Branch with paper.md** (empty if default branch):
**Version:** 1.15.0
