### 2024-07-23 82cf27e
Added random truncation method.
