+++ pref_voting/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '1.13.21'
+__version__ = '1.13.25'
