### 2024-07-30 voting-tools/pref_voting#106 @wesholliday
Hi Diane,

Thanks for your message. You can use instant_runoff_for_truncated_linear_orders to analyze the Burlington election. You can see an example here:

https://github.com/voting-tools/election-analysis/blob/main/burlington_vt_2009.ipynb

Fairly soon we will implement versions of instant runoff for an arbitrary ProfileWithTies (allowing ties in the middle of a ballot, which is not allowed in real political elections), which has been a topic of recent research.

Please let us know if you have any other questions!

Best,
Wes
