### 2024-07-30 voting-tools/pref_voting#106 @wesholliday
P.S. To clarify one point, it's true that in the Burlington election, some voters did assign multiple candidates the same rank. However, election officials consider that a mistake--an "overvote"--rather than a legitimate way of ranking candidates that the IRV algorithm will use to compute the winner. So to analyze the Burlington election at https://github.com/voting-tools/election-analysis/blob/main/burlington_vt_2009.ipynb, we first process the ballots so that all legitimate ballots have the form of truncated linear orders (i.e., there are no ties, but some candidates can be unranked; all unranked candidates are in effect considered tied for last place, which is the effect of prof.use_extended_strict_preference()).

Different jurisdictions have different rules for how to handle overvotes (see, e.g., https://github.com/voting-tools/election-analysis/blob/main/alaska_2022.ipynb) and other ballot anomalies, and unfortunately it is not always easy to find the exact rules.
