### 2024-07-31 voting-tools/pref_voting#106 @dmnapolitano
Hi Wes,

Thanks, this is all really helpful!  I can confirm that `instant_runoff_for_truncated_linear_orders` and the examples in the notebook work here (which makes sense -- the 2006 and 2009 elections are pretty similar).

Based on [this](https://pref-voting.readthedocs.io/en/latest/collective_decision_procedures.html), I assumed all profiles/graphs listed in #1 worked with all rules/methods listed there.  In hindsight, this was a silly assumption, but could I suggest you add a sentence or two to clarify that?

I'll close this issue since you're already working on it / it was my confusion 🙂
