### 2024-08-04 voting-tools/pref_voting#106 @wesholliday
Hi Diane,

Yes, we will definitely clarify this--thanks!

In the meantime, if you look at the documentation for a specific voting method, it should say under "PARAMETERS" what kind of "edata" the method accepts.

Best,
Wes
