+++ setup.py
@@ -1,12 +1,12 @@
 import setuptools
-from setuptools import   find_packages
+from setuptools import find_packages
 
 with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
     name="pref_voting",                   
-    version="1.13.27",     
+    version="1.13.28",     
     author="Eric Pacuit",
     author_email='epacuit@umd.edu',      
     description="pref_voting is a Python packaging that contains tools to reason about election profiles and margin graphs, and implementations of a variety of preferential voting methods.",
