### 2024-08-22 e6a0a99
Added functions to turn an SPO into a DiGraph or (when a linear order) a list.
