### 2024-08-22 fc82b6c
Added Ranked Pairs Defeat using a tie-breaking ordering.
