### 2024-08-25 d228045
Added funtions for enumerating margin graphs allowing ties and zero margins.
