+++ pref_voting/generate_weighted_majority_graphs.py
@@ -11,6 +11,7 @@
 
 import networkx as nx
 from itertools import combinations
+from helper import sublists, compositions, enumerate_compositions, convex_lexicographic_sublists
 from pref_voting.weighted_majority_graphs import MarginGraph
 import numpy as np
 from scipy.stats import multivariate_normal
@@ -220,7 +221,7 @@ def generate_edge_ordered_tournament_infinite_limit(num_candidates, cov_matrix =
     
     return MarginGraph(candidates, w_edges)
 
-## Generating Canonical MarginGraphs
+## Generating Canonical MarginGraphs without Tied Margins
 
 def _enumerate_ceots(num_cands, num_edges, partial_ceot, used_nodes, next_node):
 
@@ -298,7 +299,7 @@ def enumerate_canonical_edge_ordered_tournaments(num_cands, parity = "even"):
     Returns:
         A generator of ``MarginGraph`` for ``num_candidates``
         
-    .. warning:: It is only feasible to run this function for up to 5 candidates.    
+    .. warning:: It is only feasible to finish the enumeration for up to 5 candidates.    
 
     """
     
@@ -309,26 +310,10 @@ def enumerate_canonical_edge_ordered_tournaments(num_cands, parity = "even"):
                           [(e[0], e[1], 2 * (eidx + 1) if parity == "even" else 2 * eidx + 1) 
                            for eidx, e in enumerate(reversed(ceot))])
 
-def sublists(lst, length, x = None, partial_sublist = None): 
-    
-    x = length if x is None else x
-    
-    partial_sublist = list() if partial_sublist is None else partial_sublist
-    
-    if len(partial_sublist) == length: 
-        yield partial_sublist
-        
-    for i,el in enumerate(lst):
-        
-        if i < x: 
-            
-            extended_partial_sublist = partial_sublist + [el]
-            x += 1
-            yield from sublists(lst[i+1::], length, x, extended_partial_sublist)
             
 def enumerate_uniquely_weighted_margin_graphs(num_cands, weight_domain):  
     """
-    Enumerate all representatives from isomorphism classes of margin graphs with weights drawn from ``weight_domain``. 
+    Enumerate all representatives from isomorphism classes of uniquely-weighted margin graphs with weights drawn from ``weight_domain``. 
     
     Args:
         num_cands (int): the number of candidates
@@ -338,7 +323,7 @@ def enumerate_uniquely_weighted_margin_graphs(num_cands, weight_domain):
         A generator of ``MarginGraph`` for ``num_candidates``
         
 
-    .. warning:: It is only feasible to run this function for up to 5 candidates.    
+    .. warning:: It is only feasible to finish the enumeration for up to 5 candidates.    
 
     """
     
@@ -352,3 +337,140 @@ def enumerate_uniquely_weighted_margin_graphs(num_cands, weight_domain):
             yield MarginGraph(list(range(num_cands)), 
                               [(e[0], e[1], weight_list[eidx]) for eidx, e in enumerate(reversed(ceot))])
 
+
+## Generating Canonical MarginGraphs with Tied Margins
+
+def _enumerate_cweots_as_edgelist(num_cands, include_weak_tournaments=True):
+    
+    #Enumerate each canonical weakly edge ordered tournament as a list of lists of tied edges.
+    #If include_weak_tournaments = True, then allow weak tournaments in which two nodes may have no edge between them.   
+    
+
+    def edge_match(e1, e2):
+        return e1['weight'] == e2['weight']
+
+    cweots = list()
+    cweots_as_graphs = list()
+
+    for ceot in tqdm(list(_enumerate_ceots_as_edgelist(num_cands))):
+        
+        # Given ceot, we will generate many cweots as follows:
+
[…]
