### 2024-08-25 b76108c
Use invariants to speed up isomorphism checking in enumeration of canonical weakly edge-ordered tournaments
