### 2024-08-26 #7020 @dmnapolitano
Hi @britta-wstnr , I believe I'm done with my review 🎉  Sorry it took so long, and please let me know if there's anything I missed or anything else I can help out with here.

One question I have...it's minor but the authors use "etc." quite a bit throughout the paper, which is rather informal compared to the rest of the paper (which is very well-written!).  Is this ok for JOSS?  This is my first time reviewing here so I'm not entirely familiar with the style standards 🙂

Thanks!!
