### 2024-09-03 #7020 @dmnapolitano
Hi @britta-wstnr , thanks for following up!  Yes, I'm happy with how the issue was handled and I hope it'll be clarified in the documentation soon if it wasn't already.  My issue will hopefully help others, at the very least 😄

And thanks, I agree that "etc." feels out of place in this otherwise well-written paper! 🙌🏻
