### 2024-09-05 #7020 @pkrafft
Hi all and @britta-wstnr, for minor issues, should I make a list in an issue comment, or make direct edits/suggestions through a fork and pull requests?
