### 2024-09-06 #7020 @britta-wstnr
Hi @pkrafft - both works, whatever you find more convenient. If it's a list of minor comments, you can even post that in the thread here if you prefer. Bigger things we think is better documented at the software-level, i.e., on the Github repo of the software (that makes it easier for people to find it back in the future). Then just make sure to link the issue or PR here.
