### 2024-09-22 27eeee5
add binomial confidence interval
