+++ pref_voting/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '1.13.30'
+__version__ = '1.13.31'
