### 2024-09-27 2656410
add candidate types (e.g., party affiliation) to spatial profiles
