### 2024-09-27 #7020 @pkrafft
The issues I've identified in my review are as follows. I think the package, documentation, and write-up are okay but with room for improvement. They are certainly above the standard I would expect for research code but don't quite meet a high open source/free software standard.

- This might be basic but for a non-expert it could be helpful to include in the installation instructions that pip3 has to be used for Python 3 if you haven't changed your soft links from the default on some machines
- When trying to import generate_profile, I got an error "ModuleNotFoundError: No module named 'seaborn'" so there might be a problem with the dependency management in the pip file. I also get the error for preflibtools when I try to run the tests. After manually installing seaborn and preflibtools, I can run the code in the README and the tests without dependency errors.
- It would be helpful if the README included documentation on how the software tests in the tests directory are meant to be run. I have to look at the code to see they are pytest
- There is an example in the README, but not a real-world example or at least no real world context for the examples are presented in the README. As a reviewer with expertise in Python and passing knowledge of social choice theory, it's not immediately clear to me from the package write-up in the PDF nor from the README how the package is intended to be used in detail. Providing more hand-holding and guidance of real-world examples in the README would help address this. As it stands I have to dig into the Read the Docs to understand what the software is meant to do, rather than have the specific functions and capabilities be clarified by the package description. There are plenty of examples in the Read the Docs, so perhaps one or more of those could be put in the write-up and the README together with a narrative description to give a better sense for the software upfront. In particular, it takes some digging to understand what the tuples that are taken as input for the Profile objects mean.
- There is extensive functionality of the software, so it is difficult to verify all the functionality. I have been able to run the code in the README without error, but I do get three errors when I run the tests directory with pytest. This is on a fresh install with Python 3.9.6. The errors are copied below.
FAILED tests/test_all_vms.py::test_all_profile_vms - assert <networkx.cla...t 0x315056430> == [1]
FAILED tests/test_all_vms.py::test_all_profile_with_ties_vms - assert <networkx.cla...t 0x3143334f0> == [1]
FAILED tests/test_all_vms.py::test_all_margin_graph_vms - assert <networkx.cla...t 0x31457eb80> == [1]
- I did not identify any performance claims for the software
- I am not sure if the community guidelines for contributions to the software are so clear at the moment. As far as I can tell there is just an instruction to email the creators with questions.
