### 2024-09-30 95f40df
add test description to README, fix small bugs, remove setup.py, fix dependencies
