### 2024-09-30 #7020 @epacuit
Thank you for the feedback!

We are still in the process of updating the README along the lines of your suggestion.

Regarding the errors with the tests and the import issues:

1. When trying to import generate_profile, I got an error "ModuleNotFoundError: No module named 'seaborn'" so there might be a problem with the dependency management in the pip file. I also get the error for preflibtools when I try to run the tests. After manually installing seaborn and preflibtools, I can run the code in the README and the tests without dependency errors.

2. There is extensive functionality of the software, so it is difficult to verify all the functionality. I have been able to run the code in the README without error, but I do get three errors when I run the tests directory with pytest. This is on a fresh install with Python 3.9.6. The errors are copied below.
FAILED tests/test_all_vms.py::test_all_profile_vms - assert <networkx.cla...t 0x315056430> == [1]
FAILED tests/test_all_vms.py::test_all_profile_with_ties_vms - assert <networkx.cla...t 0x3143334f0> == [1]
FAILED tests/test_all_vms.py::test_all_margin_graph_vms - assert <networkx.cla...t 0x31457eb80> == [1]

These are both fixed in the commits 8856e72c890d8d7d2b02e856fac499d13c91cd96 and  95f40df3841d1f19419cfc7e23af22553348674f
