### 2024-09-30 #7020 @epacuit
Regarding the community guidelines (available here: https://github.com/voting-tools/pref_voting/blob/main/CONTRIBUTING.md): Since this is a small project, we followed the model of a related project: https://github.com/martinlackner/abcvoting/blob/master/CONTRIBUTING.md (see https://joss.theoj.org/papers/10.21105/joss.04880).

This seems the most appropriate for our project.  However, if JOSS submissions require something more extensive, please let us know.
