### 2024-10-02 #7020 @britta-wstnr
Hi @epacuit and @pkrafft,
First of all, great to see that the review process is in full swing! 🙂
Regarding your question, JOSS' guidlines can be found here: https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines
I think it would be great if you could explicitly add those three points into your Contributor Guidelines, especially "finding a bug" and "seeking support"; the contribution of new features is covered by what you already have.
Let me know if there are any questions regarding this!
