+++ pref_voting/analysis.py
@@ -9,7 +9,7 @@
 
 from pref_voting.generate_profiles import generate_profile
 from functools import partial
-from multiprocess import Pool, cpu_count
+from pathos.multiprocessing import ProcessingPool as Pool
 from scipy.stats import binomtest
 
 import pandas as pd
@@ -74,9 +74,10 @@ def find_profiles_with_different_winners(
     return profiles
 
 
-def record_condorcet_efficiency_data(vms, num_cands, num_voters, probmod, probmod_param, t):
 
-    prof = generate_profile(num_cands, num_voters, probmod=probmod, probmod_param=probmod_param)
+def record_condorcet_efficiency_data(vms, num_cands, num_voters, pm, t):
+
+    prof = pm(num_cands, num_voters)
     cw = prof.condorcet_winner()
 
     return {
@@ -84,14 +85,14 @@ def record_condorcet_efficiency_data(vms, num_cands, num_voters, probmod, probmo
         "cw_winner": {vm.name: cw is not None and [cw] == vm(prof) for vm in vms},
     }
 
-
 def condorcet_efficiency_data(
     vms,
     numbers_of_candidates=[3, 4, 5],
     numbers_of_voters=[4, 10, 20, 50, 100, 500, 1000],
-    probmods=["IC"],
-    probmod_params=None,
-    num_trials=10000,
+    prob_models = {"IC": lambda nc, nv: generate_profile(nc, nv)},
+    min_num_samples=1000,
+    max_num_samples=100_000,
+    min_error=0.01,
     use_parallel=True,
     num_cpus=12,
 ):
@@ -102,71 +103,88 @@ def condorcet_efficiency_data(
         vms (list(functions)): A list of voting methods,
         numbers_of_candidates (list(int), default = [3, 4, 5]): The numbers of candidates to check.
         numbers_of_voters (list(int), default = [5, 25, 50, 100]): The numbers of voters to check.
-        probmod (str, default="IC"): The probability model to be passed to the ``generate_profile`` method
-        num_trials (int, default=10000): The number of profiles to check for different winning sets.
+        probmod (dict, default="IC"): A dictionary with keys as the names of the probability models and values as functions that generate profiles.  Each function should accept a number of candidates and a number of voters. 
+        min_num_trials (int, default=1000): The minimum number of profiles to check.
+        max_num_trials (int, default=100_000): The maximum number of profiles to check.
+        min_error (float, default=0.01): The minimum error to allow in the 95% confidence interval.
         use_parallel (bool, default=True): If True, then use parallel processing.
         num_cpus (int, default=12): The number of (virtual) cpus to use if using parallel processing.
 
     """
 
-    probmod_params_list = [None]*len(probmods) if probmod_params is None else probmod_params
-
-    assert len(probmod_params_list) == len(probmods), "probmod_params must be a list of the same length as probmods"
-
     if use_parallel:
         pool = Pool(num_cpus)
 
     data_for_df = {
-        "vm": list(),
-        "num_cands": list(),
-        "num_voters": list(),
-        "probmod": list(),
-        "probmod_param":list(),
-        "num_trials": list(),
-        "percent_condorcet_winners": list(),
-        "condorcet_efficiency": list(),
+         "num_candidates": [],
+         "num_voters": [],
+         "prob_model": [],
+         "voting_method": [],
+         "condorcet_efficiency": [],
+         "error": [],
+         "num_samples": [],
+         "percent_condorcet_winner": [],
+         "percent_condorcet_winner_error": [],
+         "min_num_samples": list(),
+         "max_num_samples": list(),
+         "min_error": list(),
     }
-    for probmod,probmod_param in zip(probmods, probmod_params_list):
[…]
