+++ pyproject.toml
@@ -7,7 +7,7 @@ build-backend = "setuptools.build_meta"
 
 [tool.poetry]
 name = "pref_voting"
-version = "1.14.1"
+version = "1.14.2"
 description = "pref_voting is a Python package that contains tools to reason about elections and margin graphs, and implementations of voting methods."
 authors = ["Eric Pacuit <epacuit@umd.edu>"]
 license = "MIT"
@@ -29,5 +29,6 @@ prefsampling = "^0.1.16"
 filelock = "^3.12.2"
 preflibtools = "^2.0.22"
 seaborn = "^0.13.2"
+pathos = "^0.3.3"
 
 [tool.poetry.dev-dependencies]
