### 2024-10-18 881ed3e
add howto notebooks
