+++ pref_voting/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '1.14.2'
+__version__ = '1.14.4'
