### 2024-10-18 8e300ae
update README and CONTRIBUTE, and the installation guide.
