+++ CONTRIBUTING.md
@@ -1,10 +1,20 @@
 # Welcome to pref_voting!
 
-This is a small project, so the best way to contribute is to simply get in contact with Wes Holliday (wesholliday@berkeley.edu) and Eric Pacuit (epacuit@umd.edu) or to post an [issue](https://github.com/voting-tools/pref_voting/issues).
+This is a small project, so if you have any questions about the project or how to contribute, then the best thing to do is get in contact with Wes Holliday (wesholliday@berkeley.edu) and Eric Pacuit (epacuit@umd.edu).  
 
-Further comments:
-- Documentation is available at https://pref-voting.readthedocs.io/.
+## How to contribute
+
 - If you would like to contribute but don't have a concrete idea, please
 contact us directly (Wes Holliday, wesholliday@berkeley.edu and Eric Pacuit, epacuit@umd.edu).
-- If there is a voting method that you think should be included in the package, please let us know.  
+- If there is a voting method that you think should be included in the package, please let us know by posting an [issue](https://github.com/voting-tools/pref_voting/issues).  
 - If you have code that you would like to contribute, feel free to clone the repository and submit a pull request (PR) once your version is functional. We can then review and discuss your code. You're also welcome to contact us before submitting a PR.
+
+## Report issues or problems with the software
+
+- If you encounter any issues or bugs, please post an [issue](https://github.com/voting-tools/pref_voting/issues)
+
+## Seek support 
+
+- There is extensive documentation available at https://pref-voting.readthedocs.io/.
+- You can explore some of the voting methods implemented in the package by vising [https://pref.tools/pref_voting](https://pref.tools/pref_voting).
+- There are a number of helpful notebooks in the [howto directory](https://github.com/voting-tools/pref_voting/tree/main/howto) of the repository that demonstrate how to use the package.
