+++ docs/source/installation.rst
@@ -1,10 +1,21 @@
 Installation
 ================
 
-
 ``pref_voting`` is available on pypi and can be installed using pip.
 
 .. code-block:: shell
 
-    pip install pref_voting
+    pip3 install pref_voting
+
+Alternatively, you can install the latest development version from source:
+
+.. code-block:: shell
+
+    git clone https://github.com/voting-tools/pref_voting
+    pip3 install .
+
+Requirements:
+
+* Python 3.9+
+* The following pip packages are required and installed automatically:  filelock, scipy, networkx, tabulate, matplotlib, nashpy, numba, ortools, pathos, preflibtools, prefsampling, and seaborn 
 
