### 2024-10-24 99ec375
Added to_latex method
