### 2024-10-24 dbed236
update README, fix documentation, document PairwiseComparisons
