+++ docs/source/pairwise_profiles.md
@@ -1,0 +1,27 @@
+Pairwise Profiles
+=======================================
+
+
+## Pairwise Profile Class
+
+
+```{eval-rst}
+
+
+.. autoclass:: pref_voting.pairwise_profiles.PairwiseComparisons
+    :members: 
+
+
+```
+
+
+```{eval-rst}
+
+
+.. autoclass:: pref_voting.pairwise_profiles.PairwiseProfile
+    :members: 
+
+
+```
+
+
