### 2024-10-24 734277c
fix rp_defeat return bug
