### 2024-10-27 af936cb
update numba version
