+++ pyproject.toml
@@ -7,7 +7,7 @@ build-backend = "setuptools.build_meta"
 
 [tool.poetry]
 name = "pref_voting"
-version = "1.14.6"
+version = "1.14.7"
 description = "pref_voting is a Python package that contains tools to reason about elections and margin graphs, and implementations of voting methods."
 authors = ["Eric Pacuit <epacuit@umd.edu>"]
 license = "MIT"
@@ -23,7 +23,7 @@ tabulate = "^0.9.0"
 random2 = "^1.0.1"
 matplotlib = "^3.5.2"
 nashpy = "^0.0.40"
-numba = "^0.58.0"
+numba = "^0.60.0"
 ortools = "^9.8.0"
 prefsampling = "^0.1.16"
 filelock = "^3.12.2"
