### 2024-10-28 20f3521
fix issue in Profile.description  resulting for numpy upgrade
