### 2024-11-08 60a6a87
Added method for applying candidate permutation to profile.
