### 2024-11-08 511fc4b
Added tests for apply_cand_permutation
