### 2024-11-10 e1466bf
fix profile tests
