### 2024-11-21 3ff15b5
Added Plurality Veto method from https://arxiv.org/abs/2305.19632, as well as test of apply_cand_permutation
