### 2024-11-21 3a32f04
Added warning to documentation of plurality_with_veto
