+++ pref_voting/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '1.14.9'
+__version__ = '1.14.11'
