### 2024-11-22 d868ef8
Added tests of plurality_veto
