### 2024-11-22 2c7952f
Added function for taking the mixture of two probabilistic voting methods
