### 2024-11-22 20d4d4e
Added RaDiUS method and fixed random_consensus builder
