### 2024-11-22 205f17e
Update probabilistic_methods.py
