### 2024-11-22 16fdde7
Added some stochastic voting methods and fixed floating point artifacts for Maximal Lotteries
