+++ docs/source/stochastic_methods.md
@@ -1,0 +1,26 @@
+Stochastic Methods
+=======================================
+
+## Random Consensus Builder (Stochastic)
+
+```{eval-rst}
+
+.. autofunction:: pref_voting.probabilistic_methods.random_consensus_builder_st
+
+```
+
+## Maximal Lotteries mixed with Random Consensus Builder
+
+```{eval-rst}
+
+.. autofunction:: pref_voting.probabilistic_methods.MLRCB
+
+```
+
+## Maximal Lotteries mixed with RaDiUS
+
+```{eval-rst}
+
+.. autofunction:: pref_voting.probabilistic_methods.MLRaDiUS
+
+```
