+++ pref_voting/probabilistic_methods.py
@@ -158,6 +158,9 @@ def random_consensus_builder(profile, curr_cands=None, beta=0.5):
 
     Returns:
         dict: Maps each candidate to their probability of winning under the RCB method
+
+    .. seealso::
+        :meth:`pref_voting.stochastic_methods.random_consensus_builder_st`
 """
     if curr_cands is None:
         curr_cands = profile.candidates
