### 2024-11-23 949f124
Added new iterative methods to docs
