### 2024-11-24 5dba0e5
add lp algorithm to find maximal lotteries
