+++ pref_voting/probabilistic_methods.py
@@ -8,7 +8,7 @@
 
 from pref_voting.prob_voting_method import  *
 from pref_voting.weighted_majority_graphs import  MajorityGraph, MarginGraph
-
+from scipy.optimize import linprog
 
 import random
 import nashpy as nash
@@ -70,8 +70,17 @@ def pr_borda(profile, curr_cands=None):
 
     return {c: borda_scores[c] / total_borda_scores for c in borda_scores.keys()}
 
+def clean_and_normalize(probs, threshold=1e-10):
+    # Set  negative or small positive values to zero
+    probs = np.where(probs < threshold, 0, probs)
+    
+    # Renormalize to ensure the probabilities sum to 1
+    total = np.sum(probs)
+    if total > 0:
+        probs /= total
+    return probs
 
-def _maximal_lottery(edata, curr_cands=None, margin_transformation=lambda x: x):
+def _maximal_lottery_enumeration(edata, curr_cands=None, margin_transformation=lambda x: x):
     '''
     Implementation of maximal lotteries. See http://dss.in.tum.de/files/brandt-research/fishburn_slides.pdf 
     
@@ -85,33 +94,89 @@ def _maximal_lottery(edata, curr_cands=None, margin_transformation=lambda x: x):
 
     # Create the game
     game = nash.Game(A)
-    # Find the Nash Equilibrium with Vertex Enumeration
-    equilibria = list(game.vertex_enumeration())
+    equilibria = []
+    try:
+        equilibria = list(game.vertex_enumeration())
+        #print("Vertex Enumeration found equilibria.")
+    except Exception as e:
+        print(f"Vertex Enumeration failed: {e}")
+
+    # Backup method 1: Support Enumeration
+    if not equilibria:
+        try:
+            equilibria = list(game.support_enumeration())
+            #print("Support Enumeration found equilibria.")
+        except Exception as e:
+            print(f"Support Enumeration failed: {e}")
+
     if len(equilibria) == 0:
         return {c: 1 / len(candidates) for c in candidates}
     else:
-        eq = random.choice(equilibria)
-        eq_probs = eq[0]
+        # average the  equilibria component-wise to get a single equilibrium
+        eq_probs = [np.mean([eq[0][idx] for eq in equilibria]) 
+                    for idx in range(len(candidates))]
 
-        # Clip small negative values and normalize
-        eq_probs = np.maximum(eq_probs, 0)  # Remove floating-point artifacts
-        if not np.isclose(sum(eq_probs), 1):  # Normalize if necessary
-            eq_probs /= sum(eq_probs)
-
+        eq_probs = clean_and_normalize(np.array(eq_probs))
+        
         # Return the result as a dictionary
         return {c: eq_probs[cand_to_cidx(c)] for c in candidates}
 
+
+def _maximal_lottery_lp(edata, curr_cands=None, margin_transformation=lambda x: x):
+    '''
+    Implementation of maximal lotteries using linear programming.
+    '''
+    candidates = edata.candidates if curr_cands is None else curr_cands
+    m_matrix, cand_to_cidx = edata.strength_matrix(curr_cands=candidates)
+
+    A = np.array([[margin_transformation(m) for m in row] for row in m_matrix])
+
+    num_cands = len(candidates)
+    c = np.zeros(num_cands + 1)
+    c[-1] = -1  # Coefficient for v in the objective function (maximize v)
+
+    # Inequalities: A^T p - v * 1 >= 0 => A^T p - v * 1 - s = 0 (s >= 0)
+    # We need to convert this to the form: A_ub x <= b_ub
+
+    A_ub = np.hstack([-A.T, np.ones((num_cands, 1))])
+    b_ub = np.zeros(num_cands)
+
[…]
