### 2024-11-27 17c7fd1
clean up notebooks; fix license link
