+++ pref_voting/voting_methods.py
@@ -4,4 +4,5 @@ from pref_voting.c1_methods import *
 from pref_voting.margin_based_methods import *
 from pref_voting.combined_methods import *
 from pref_voting.other_methods import *
+from pref_voting.stochastic_methods import *
 
