### 2024-12-01 38a337a
Updated docstrings concerning RCB
