### 2024-12-07 da6e4da
Updated variable name in Condorcet efficiency function
