### 2024-12-09 #7020 @britta-wstnr
@pkrafft If the examples are not working on MacOS, do you think this could be fixed by @epacuit ? i.e., can you provide pointers on what is going wrong? Or do you think it's a very specific problem that others will not run into?
