### 2024-12-09 #7020 @pkrafft
Hi both, I'm not sure. I'm a bit stuck because I'd have to reformat my machine to be able to confidently reproduce the error. I think what happened was I installed the package with native python and pip when I did my initial review and then in the interim I had to set up a new python with Brew to do some other work, which then broke my old pip. So now Brew can't find some of the package dependencies, and it can't install them either---I think because they aren't registered on Brew? What I am mostly not sure about though is if I had done everything from the start from a fresh slate just with Brew if I would have this problem. That's the main case I think could potentially cause trouble for others. Brew is one of the standard python package managers for Mac OS. Hope that's clear?
