### 2024-12-09 #7020 @epacuit
Thanks!  I'm happy to help if I can, but I'm not completely sure that I understand the problem.  From what I understand (and I definitely could be wrong!) is that brew is a MacOS package manager and pip/pip3 is the Python package manager.    I'm not sure that one would install most Python packages using brew.   But, perhaps I am misunderstanding the problem?

 I'm certainly happy to make it easier for people to install pref_voting, but I do worry that not all the required packages will be available to be installed using brew (e.g., nashpy, preflibtools, and prefsampling).   Is it required for the JOSS publication to have the package (and all dependences) be able to be installed using brew as well as pip?

 Perhaps you can let me know which package dependences can't be found using brew?  Definitely, pref_voting and the dependencies  nashpy, preflibtools, and prefsampling aren't registered on brew, but they can all be installed using pip.

 Does it work to create a virtual environment and do a "pip3 install pref_voting"?   This should install all the required packages.
