### 2024-12-12 #7020 @pkrafft
Hi both,

Thanks for that helpful tip, @epacuit. You're right it seems the virtual environment ignores Brew, so that's great. For your reference, and if anyone searches the error message online, the error I was previously getting was as follows.

>
> % pip3 install pref_voting
>
> [notice] A new release of pip is available: 24.2 -> 24.3.1
> [notice] To update, run: python3.13 -m pip install --upgrade pip
> error: externally-managed-environment
>
> × This environment is externally managed
> ╰─> To install Python packages system-wide, try brew install
>     xyz, where xyz is the package you are trying to
>     install.
>
>     If you wish to install a Python library that isn't in Homebrew,
>     use a virtual environment:
>
>     python3 -m venv path/to/venv
>     source path/to/venv/bin/activate
>     python3 -m pip install xyz
>
>     If you wish to install a Python application that isn't in Homebrew,
>     it may be easiest to use 'pipx install xyz', which will manage a
>     virtual environment for you. You can install pipx with
>
>     brew install pipx
>
>     You may restore the old behavior of pip by passing
>     the '--break-system-packages' flag to pip, or by adding
>     'break-system-packages = true' to your pip.conf file. The latter
>     will permanently disable this error.
>
>     If you disable this error, we STRONGLY recommend that you additionally
>     pass the '--user' flag to pip, or set 'user = true' in your pip.conf
>     file. Failure to do this can result in a broken Homebrew installation.
>
>     Read more about this behavior here: <https://peps.python.org/pep-0668/>
>
> note: If you believe this is a mistake, please contact your Python installation or OS distribution provider. You can override this, at the risk of breaking your Python installation or OS, by passing --break-system-packages.
> hint: See PEP 668 for the detailed specification.

So I used, "% brew install virtualenv", and then opened a virtual environment, and the above worked in that environment.

**However, I ran into a subsequent error:**

> Collecting numba<0.61.0,>=0.60.0 (from pref_voting)
>   Downloading numba-0.60.0.tar.gz (2.7 MB)
>      ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 2.7/2.7 MB 21.4 MB/s eta 0:00:00
>   Installing build dependencies ... done
>   Getting requirements to build wheel ... error
>   error: subprocess-exited-with-error
>
>   × Getting requirements to build wheel did not run successfully.
>   │ exit code: 1
>   ╰─> [24 lines of output]
>       Traceback (most recent call last):
>         File "/Users/pkrafft/Documents/Reviewing/pref_voting/venv/lib/python3.13/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 353, in <module>
>           main()
>           ~~~~^^
>         File "/Users/pkrafft/Documents/Reviewing/pref_voting/venv/lib/python3.13/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 335, in main
>           json_out['return_val'] = hook(**hook_input['kwargs'])
>                                    ~~~~^^^^^^^^^^^^^^^^^^^^^^^^
>         File "/Users/pkrafft/Documents/Reviewing/pref_voting/venv/lib/python3.13/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 118, in get_requires_for_build_wheel
>           return hook(config_settings)
>         File "/private/var/folders/tn/zr42l_zj4yg_lmfz83d7_fkc0000gr/T/pip-build-env-xqmc2lhs/overlay/lib/python3.13/site-packages/setuptools/build_meta.py", line 334, in get_requires_for_build_wheel
>           return self._get_build_requires(config_settings, requirements=[])
>                  ~~~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
>         File "/private/var/folders/tn/zr42l_zj4yg_lmfz83d7_fkc0000gr/T/pip-build-env-xqmc2lhs/overlay/lib/python3.13/site-packages/setuptools/build_meta.py", line 304, in _get_build_requires
>           self.run_setup()
>           ~~~~~~~~~~~~~~^^
>         File "/private/var/folders/tn/zr42l_zj4yg_lmfz83d7_fkc0000gr/T/pip-build-env-xqmc2lhs/overlay/lib/python3.13/site-packages/setuptools/build_meta.py", line 522, in run_setup
>           super().run_setup(setup_script=setup_script)
>           ~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^
>         File "/private/var/folders/tn/zr42l_zj4yg_lmfz83d7_fkc0000gr/T/pip-build-env-xqmc2lhs/overlay/lib/python3.13/site-packages/setuptools/build_meta.py", line 320, in run_setup
>           exec(code, locals())
>           ~~~~^^^^^^^^^^^^^^^^
>         File "<string>", line 51, in <module>
>         File "<string>", line 48, in _guard_py_ver
>       RuntimeError: Cannot install on Python version 3.13.0; only versions >=3.9,<3.13 are supported.
>       [end of output]
>
>   note: This error originates from a subprocess, and is likely not a problem with pip.
> error: subprocess-exited-with-error
>
> × Getting requirements to build wheel did not run successfully.
> │ exit code: 1
> ╰─> See above for output.
>
> note: This error originates from a subprocess, and is likely not a problem with pip.

**So I changed the virual environment to use Python 3.12, and that resolved that error.

But then I encountered another one that I am not sure how to fix:**

**Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/pkrafft/Documents/Reviewing/pref_voting/pref_voting/profiles.py", line 19, in <module>
    from pref_voting.voting_method import _num_rank_first
  File "/Users/pkrafft/Documents/Reviewing/pref_voting/pref_voting/voting_method.py", line 18, in <module>
    import pkg_resources
ModuleNotFoundError: No module named 'pkg_resources'**
