### 2024-12-12 99e62b1
remove pkg_resources, add importlib; add numba warning to installation guide
