+++ docs/source/conf.py
@@ -18,11 +18,11 @@
 # -- Project information -----------------------------------------------------
 
 project = 'pref_voting'
-copyright = '2023, Wes Holliday and Eric Pacuit'
+copyright = '2024, Wes Holliday and Eric Pacuit'
 author = 'Wes Holliday and Eric Pacuit'
 
 # The full version, including alpha/beta/rc tags
-release = '1.14.18'
+release = '1.14.19'
 
 
 # -- General configuration ---------------------------------------------------
