### 2024-12-12 50514eb
fix link in installation.rst
