### 2024-12-12 #7020 @epacuit
Thanks!   I see the issue:   pkg_resources is deprecated in Python 3.12.

I just updated to the latest version (1.14.19) that removes pkg_resources and uses importlib.  So, if you install the latest version of pref_voting it should work.

Regarding the incompatibility with Python 3.13.     The issue here is that we are using Numba, and so we have to wait until Numba releases a version that is compatible with Python 3.13.   It usually takes a few months (sometimes longer) for Numba to release a version that is compatible with the latest version of Python.   I added a comment to the README and the installation guide in the documentation about this.  Both of them point people to the Numba documentation about which versions of Python (and numpy) are compatible with Numba: https://numba.readthedocs.io/en/stable/user/installing.html#version-support-information.   I'm not sure of the best practice for addressing this issue, so I'm happy to accept any other suggestions.
