### 2024-12-13 #7020 @pkrafft
Thanks, @epacuit!

I've now been able to successfully run the examples, but getting a similar error in the tests.

% pytest
ImportError while loading conftest '/Users/pkrafft/Documents/Reviewing/pref_voting/tests/conftest.py'.
conftest.py:3: in <module>
    from pref_voting.profiles import Profile
../pref_voting/profiles.py:14: in <module>
    from numba import jit
E   ModuleNotFoundError: No module named 'numba'

I'm not sure why this is happening because when I open a python command line and use "from numba import jit" within the same directory and virtual environment, it loads properly.

As for best practices on the documentation, @britta-wstnr would probably have better suggestsions, but you might at least note prominently what the last version of Python you've tested the software, if not making that version a requirement to run it.
