### 2024-12-13 #7020 @epacuit
Hmm, strange.  My guess is that pytest is using a different version of Python.     Perhaps using "python -m pytest" or "python3 -m pytest" is solves the problem?
