### 2024-12-13 547c1ca
add comment about Python 3.12 to README
