### 2024-12-13 #7020 @epacuit
I also added a comment to the README that the package has been tested on Python 3.12
