### 2024-12-13 #7020 @pkrafft
Oh, and just to say explicitly, the last solution was to reinstall pytest in the virtual environment.
