### 2024-12-26 3b7e602
Corrected documentation for axiom_violation_data
