+++ pref_voting/analysis.py
@@ -303,14 +303,14 @@ def axiom_violations_data(
     num_cpus=12,
 ):
     """
-    Returns a Pandas DataFrame with the Condorcet efficiency of a list of voting methods.
+    Returns a Pandas DataFrame with axiom violation data for a list of voting methods.
 
     Args:
         vms (list(functions)): A list of voting methods,
         numbers_of_candidates (list(int), default = [3, 4, 5]): The numbers of candidates to check.
         numbers_of_voters (list(int), default = [5, 25, 50, 100]): The numbers of voters to check.
         probmod (str, default="IC"): The probability model to be passed to the ``generate_profile`` method
-        num_trials (int, default=10000): The number of profiles to check for different winning sets.
+        num_trials (int, default=10000): The number of profiles to check for axiom violations.
         use_parallel (bool, default=True): If True, then use parallel processing.
         num_cpus (int, default=12): The number of (virtual) cpus to use if using parallel processing.
 
@@ -489,99 +489,3 @@ def means_with_estimated_standard_error(
             mean_for_each_experiment=means)
 
     return means, est_std_errors, variances, num_trials
-
-
-#### Bootstrap confidence interval analysis (to be removed) ####
-
-def bootstrap_cia(
-        generate_samples,
-        process_samples, 
-        precision, 
-        averaging_axis = 0,
-        confidence=0.95, 
-        initial_trials=10000, 
-        step_trials=1000, 
-        bootstrap_samples=1000,
-        verbose = False, 
-        max_num_trials = None
-        ):
-    """
-    Applies the percentile bootstrap confidence interval analysis using the functions generate_samples and process_samples.
-    
-    Args:
-        generate_samples (function): A function that generates samples.
-        process_samples (function): A function that processes the samples generated by generate_samples.
-        precision (float): The desired precision of the confidence interval.
-        averaging_axis (int, default=0): The axis along which to average the samples.
-        confidence (float, default=0.95): The desired confidence level of the confidence interval.
-        initial_trials (int, default=10000): The number of samples to initially generate.
-        step_trials (int, default=1000): The number of samples to generate in each step.
-        bootstrap_samples (int, default=1000): The number of bootstrap samples to select.
-        verbose (bool, default=False): If True, then print progress information.
-        max_num_trials (int, default=None): If not None, then the maximum number of trials to run.
-
-    Returns:
-        A tuple (means, half_widths, variance_bootstrap_means, variance_of_values, num_trials) where means is an array of the means of the bootstrap means, half_widths is an array of the half widths of the confidence intervals, variance_bootstrap_means is an array of the variances of the bootstrap means, and variance_of_values is an array of the variances of the values returned by process_samples, and num_trials is the total number of trials.
-
-    """
-    
-    num_gen_samples = 0
-    # Generate initial samples
-    samples = generate_samples(num_samples = initial_trials, num_gen_samples = num_gen_samples)
-
-    # Process the results
-    values = process_samples(samples)
-
-    # Generate bootstrap samples and calculate the confidence interval for values
-    bootstrap_means = np.array([
-        [np.mean(np.random.choice(s, size=len(s))) for s in values] 
-        for _ in range(bootstrap_samples)
-    ])
-    
-    means = np.mean(bootstrap_means, axis=0)
-
-    lowers, uppers = np.percentile(bootstrap_means, 
-                                     [(1-confidence)/2*100, 
-                                      (1+confidence)/2*100], axis=0)
-    half_widths = (uppers - lowers) / 2
-    
-    variance_bootstrap_means = np.var(bootstrap_means, axis=0)
-    
-    # Compute variance of each array in values
[…]
