### 2024-12-26 77ca4f1
Updated resoluteness_data and added to docs
