+++ docs/source/analysis_overview.md
@@ -26,6 +26,17 @@ Functions that can be used to compare and contrast voting methods.
 
 ```
 
+## Resoluteness
+
+
+```{eval-rst}
+
+
+.. autofunction:: pref_voting.analysis.resoluteness_data
+
+
+```
+
 ## Axiom Violations  
 
 ```{eval-rst}
