### 2025-01-01 7d22f38
Added bucklin_for_truncated_linear_orders
