### 2025-01-07 c4f4066
Adding missing return to single_voter_resolvability
