+++ pref_voting/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '1.14.20'
+__version__ = '1.14.22'
