+++ README.md
@@ -19,6 +19,7 @@ pip3 install pref_voting
 ```
 **Notes**: 
 * The package requires Python 3.9 or higher and has been tested on Python 3.12.
+
 * Since the package uses Numba, refer to the [Numba documentation for the latest supported Python version](https://numba.readthedocs.io/en/stable/user/installing.html#version-support-information).
 * If you have both Python 2 and Python 3 installed on your system, make sure to use ``pip3`` instead of pip to install packages for Python 3. Alternatively, you can use ``python3 -m pip`` to ensure you're using the correct version of pip. If you have modified your system's defaults or soft links, adjust accordingly.
 
