### 2025-01-09 afde4f3
Added Neutral Reversal to variable voter axioms
