### 2025-01-11 bed2438
add preferential equality axiom
