+++ pref_voting/variable_voter_axioms.py
@@ -1889,6 +1889,143 @@ single_voter_resolvability = Axiom(
     find_all_violations = find_all_single_voter_resolvability_violations, 
 )
 
+
+def has_preferential_equality_violation(prof, vm, verbose=False):
+    """
+    Check if a profile has a preferential equality violation for the voting method vm.
+
+    See Definition 2.1 and Lemma 2.4 from the paper "Characterizations of voting rules based on majority margins" by Y. Ding,  W. Holliday, and E. Pacuit.
+
+    """
+    prof_constructor = ProfileWithTies if isinstance(prof, ProfileWithTies) else Profile
+    
+    for x, y in combinations(prof.candidates, 2):
+        
+        xy_rankings = [r for r in prof.rankings if get_rank(r, x) == get_rank(r, y) - 1]
+
+        other_rankings = [r for r in prof.rankings if r not in xy_rankings]
+
+        if len(xy_rankings) != 0 and len(xy_rankings) % 2 == 0:
+            for I, J in equal_size_partitions_with_duplicates(xy_rankings):
+                new_rankings_I = [swap_candidates(r, x, y) for r in I] + list(J) + list(other_rankings)
+                prof_I = prof_constructor(new_rankings_I)
+                new_rankings_J = [swap_candidates(r, x, y) for r in J] + list(I) + list(other_rankings)
+                 
+                prof_J = prof_constructor(new_rankings_J)
+
+                if vm(prof_I) != vm(prof_J): 
+                    if verbose:
+                        print("The original profile")
+                        prof.anonymize().display()
+                        print(prof.description())
+                        print(f"The profile after swapping {x} and {y} in the rankings {I}:")
+                        prof_I.anonymize().display()
+                        print(prof_I.description())
+                        vm.display(prof_I)
+                        print(f"The profile after swapping {x} and {y} in the rankings {J}:")
+                        prof_J.anonymize().display()
+                        print(prof_J.description())
+                        vm.display(prof_J)
+                    return True
+
+        yx_rankings = [r for r in prof.rankings if get_rank(r, y) == get_rank(r, x) - 1]
+
+        other_rankings = [r for r in prof.rankings if r not in yx_rankings]
+
+        if len(yx_rankings) != 0 and len(yx_rankings) % 2 == 0:
+            for I, J in equal_size_partitions_with_duplicates(yx_rankings):
+                new_rankings_I = [swap_candidates(r, y, x) for r in I] + list(J) + list(other_rankings)
+                prof_I = prof_constructor(new_rankings_I)
+                new_rankings_J = [swap_candidates(r, y, x) for r in J] + list(I) + list(other_rankings)
+                prof_J = prof_constructor(new_rankings_J)
+
+                if vm(prof_I) != vm(prof_J): 
+                    if verbose:
+                        print("The original profile")
+                        prof.anonymize().display()
+                        print(prof.description())
+                        print(f"The profile after swapping {x} and {y} in the rankings {I}:")
+                        prof_I.anonymize().display()
+                        print(prof_I.description())
+                        vm.display(prof_I)
+                        print(f"The profile after swapping {x} and {y} in the rankings {J}:")
+                        prof_J.anonymize().display()
+                        print(prof_J.description())
+                        vm.display(prof_J)
+                    return True
+    return False
+
+def find_all_preferential_equality_violations(prof, vm, verbose=False):
+    """
[…]
