+++ docs/source/conf.py
@@ -22,7 +22,7 @@ copyright = '2024, Wes Holliday and Eric Pacuit'
 author = 'Wes Holliday and Eric Pacuit'
 
 # The full version, including alpha/beta/rc tags
-release = '1.14.23'
+release = '1.14.25'
 
 
 # -- General configuration ---------------------------------------------------
