+++ pref_voting/variable_voter_axioms copy.py
@@ -1,0 +1,2155 @@
+"""
+    File: variable_voter_axioms.py
+    Author: Wes Holliday (wesholliday@berkeley.edu) and Eric Pacuit (epacuit@umd.edu)
+    Date: January 9, 2024
+    
+    Variable voter axioms 
+"""
+
+from pref_voting.axiom import Axiom
+from pref_voting.axiom_helpers import *
+import numpy as np
+from itertools import product, combinations, permutations
+from pref_voting.helper import weak_orders
+from pref_voting.rankings import Ranking
+
+def divide_electorate(prof):
+    """Given a Profile or ProfileWithTies object, yield all possible ways to divide the electorate into two nonempty electorates."""
+
+    R, C = prof.rankings_counts
+
+    ranges = [range(count+1) for count in C]
+
+    # For each combination of divisions
+    for division in product(*ranges):
+        C1 = np.array(division)
+        C2 = C - C1
+
+        # We will filter out rankings where the count is zero
+        nonzero_indices_C1 = np.nonzero(C1)[0]
+        nonzero_indices_C2 = np.nonzero(C2)[0]
+
+        # Only yield if both electorates have at least one voter
+        if nonzero_indices_C1.size > 0 and nonzero_indices_C2.size > 0:
+
+            rankings1 = R[nonzero_indices_C1].tolist()
+            rankings2 = R[nonzero_indices_C2].tolist()
+            counts1 = C1[nonzero_indices_C1].tolist()
+            counts2 = C2[nonzero_indices_C2].tolist()
+
+            if rankings1 <= rankings2: # This prevents yielding both prof1, prof2 and later on prof2, prof1, unless they are equal
+
+                if isinstance(prof,Profile):
+                    prof1 = Profile(rankings1, rcounts = counts1)
+                    prof2 = Profile(rankings2, rcounts = counts2)
+                
+                if isinstance(prof,ProfileWithTies):
+                    prof1 = ProfileWithTies(rankings1, rcounts = counts1)
+                    prof2 = ProfileWithTies(rankings2, rcounts = counts2)
+
+                    if prof.using_extended_strict_preference:
+                        prof1.use_extended_strict_preference()
+                        prof2.use_extended_strict_preference()
+            
+                yield prof1, prof2
+
+
+def has_reinforcement_violation_with_undergeneration(prof, vm, verbose=False):
+    """Returns true if there is some binary partition of the electorate such that some candidate wins in both subprofiles but not in the full profile"""
+    ws = vm(prof)
+
+    for prof1, prof2 in divide_electorate(prof):
+        winners_in_both = [c for c in vm(prof1) if c in vm(prof2)]
+        if len(winners_in_both) > 0:
+            undergenerated = [c for c in winners_in_both if c not in ws]
+            if len(undergenerated) > 0:
+                if verbose:
+                    print(f"Candidate {undergenerated[0]} wins in subprofiles 1 and 2 but loses in the full profile:")
+                    print("")
+                    print("Subprofile 1")
+                    prof1.display()
+                    print(prof1.description())
+                    vm.display(prof1)
+                    print("")
+                    print("Subprofile 2")
+                    prof2.display()
+                    print(prof2.description())
+                    vm.display(prof2)
+                    print("")
+                    print("Full profile")
+                    prof.display()
+                    print(prof.description())
+                    vm.display(prof)
+                    print("")
+                return True
+            
+    return False
+
+def has_reinforcement_violation_with_overgeneration(prof, vm, verbose=False):
+    """Returns true if there is some binary partition of the electorate such that some candidate wins in both subprofiles 
+    but there is a winner in the full profile who is not among the winners in both subprofiles"""
+
+    ws = vm(prof)
[…]
