+++ pref_voting/variable_voter_axioms.py
@@ -2008,142 +2008,62 @@ neutral_reversal = Axiom(
 )
 
 
-def has_preferential_equality_violation(prof, vm, verbose=False):
+def has_neutral_indifference_violation(prof, vm, verbose=False): 
     """
-    Check if a profile has a preferential equality violation for the voting method vm.
-
-    See Definition 2.1 and Lemma 2.4 from the paper "Characterizations of voting rules based on majority margins" by Y. Ding,  W. Holliday, and E. Pacuit.
-
+    Return True if the profile prof has a neutral indifference violation for the voting method vm.  Otherwise, return False.  That is, return True if there is a tie ranking that can be added to the profile that changes winning set according to vm.  Otherwise, return False.
     """
-    prof_constructor = ProfileWithTies if isinstance(prof, ProfileWithTies) else Profile
-    
-    for x, y in combinations(prof.candidates, 2):
-        
-        xy_rankings = [r for r in prof.rankings if get_rank(r, x) == get_rank(r, y) - 1]
 
-        other_rankings = [r for r in prof.rankings if r not in xy_rankings]
-
-        if len(xy_rankings) != 0 and len(xy_rankings) % 2 == 0:
-            for I, J in equal_size_partitions_with_duplicates(xy_rankings):
-                new_rankings_I = [swap_candidates(r, x, y) for r in I] + list(J) + list(other_rankings)
-                prof_I = prof_constructor(new_rankings_I)
-                new_rankings_J = [swap_candidates(r, x, y) for r in J] + list(I) + list(other_rankings)
-                 
-                prof_J = prof_constructor(new_rankings_J)
-
-                if vm(prof_I) != vm(prof_J): 
-                    if verbose:
-                        print("The original profile")
-                        prof.anonymize().display()
-                        print(prof.description())
-                        print(f"The profile after swapping {x} and {y} in the rankings {I}:")
-                        prof_I.anonymize().display()
-                        print(prof_I.description())
-                        vm.display(prof_I)
-                        print(f"The profile after swapping {x} and {y} in the rankings {J}:")
-                        prof_J.anonymize().display()
-                        print(prof_J.description())
-                        vm.display(prof_J)
-                    return True
-
-        yx_rankings = [r for r in prof.rankings if get_rank(r, y) == get_rank(r, x) - 1]
-
-        other_rankings = [r for r in prof.rankings if r not in yx_rankings]
-
-        if len(yx_rankings) != 0 and len(yx_rankings) % 2 == 0:
-            for I, J in equal_size_partitions_with_duplicates(yx_rankings):
-                new_rankings_I = [swap_candidates(r, y, x) for r in I] + list(J) + list(other_rankings)
-                prof_I = prof_constructor(new_rankings_I)
-                new_rankings_J = [swap_candidates(r, y, x) for r in J] + list(I) + list(other_rankings)
-                prof_J = prof_constructor(new_rankings_J)
-
-                if vm(prof_I) != vm(prof_J): 
-                    if verbose:
-                        print("The original profile")
-                        prof.anonymize().display()
-                        print(prof.description())
-                        print(f"The profile after swapping {x} and {y} in the rankings {I}:")
-                        prof_I.anonymize().display()
-                        print(prof_I.description())
-                        vm.display(prof_I)
-                        print(f"The profile after swapping {x} and {y} in the rankings {J}:")
-                        prof_J.anonymize().display()
-                        print(prof_J.description())
-                        vm.display(prof_J)
-                    return True
+    tie_ranking = Ranking({c:0 for c in prof.candidates})
+    if isinstance(prof, ProfileWithTies):
+        new_rankings = prof.rankings + [tie_ranking]
[…]
