### 2025-01-12 4f828f3
add nonlinear_neutral_reversal and neutral_indifference
