+++ pref_voting/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '1.14.26'
+__version__ = '1.14.28'
