### 2025-01-13 5b46492
Added line skips to verbose modes of axioms
