+++ pref_voting/variable_voter_axioms.py
@@ -2027,6 +2027,7 @@ def has_neutral_indifference_violation(prof, vm, verbose=False):
             prof.anonymize().display()
             print(prof.description())
             vm.display(prof)
+            print("")
             print(f"The profile after adding a tie ranking:")
             new_prof.anonymize().display()
             print(new_prof.description())
@@ -2052,6 +2053,7 @@ def find_all_neutral_indifference_violations(prof, vm, verbose=False):
             prof.anonymize().display()
             print(prof.description())
             vm.display(prof)
+            print("")
             print(f"The profile after adding a tie ranking:")
             new_prof.anonymize().display()
             print(new_prof.description())
@@ -2088,6 +2090,7 @@ def has_nonlinear_neutral_reversal_violation(prof, vm, verbose=False):
                 prof.anonymize().display()
                 print(prof.description())
                 vm.display(prof)
+                print("")
                 print(f"The profile after adding {ranking} and its reverse {ranking_reverse}:")
                 new_prof.anonymize().display()
                 print(new_prof.description())
@@ -2121,6 +2124,7 @@ def find_all_nonlinear_neutral_reversal_violations(prof, vm, verbose=False):
                 prof.anonymize().display()
                 print(prof.description())
                 vm.display(prof)
+                print("")
                 print(f"The profile after adding {ranking} and its reverse {ranking_reverse}:")
                 new_prof.anonymize().display()
                 print(new_prof.description())
