### 2025-01-13 d0a32a1
add tiebreaking compensation
