### 2025-01-13 889d7cb
fix is_tied method of a Ranking object
