### 2025-01-13 0b6b300
Updated how rankings are displayed in verbose mode of Preferential Equality
