+++ pref_voting/invariance_axioms.py
@@ -368,11 +368,11 @@ def has_preferential_equality_violation(prof, vm, verbose=False):
                         print("The original profile")
                         prof.anonymize().display()
                         print(prof.description())
-                        print(f"\nThe profile after swapping {x} and {y} in the rankings {I}:")
+                        print(f"\nThe profile after swapping {x} and {y} in the rankings {[r.rmap for r in I]}:")
                         prof_I.anonymize().display()
                         print(prof_I.description())
                         vm.display(prof_I)
-                        print(f"\nThe profile after swapping {x} and {y} in the rankings {J}:")
+                        print(f"\nThe profile after swapping {x} and {y} in the rankings {[r.rmap for r in J]}:")
                         prof_J.anonymize().display()
                         print(prof_J.description())
                         vm.display(prof_J)
@@ -394,11 +394,11 @@ def has_preferential_equality_violation(prof, vm, verbose=False):
                         print("The original profile")
                         prof.anonymize().display()
                         print(prof.description())
-                        print(f"\nThe profile after swapping {x} and {y} in the rankings {I}:")
+                        print(f"\nThe profile after swapping {x} and {y} in the rankings {[r.rmap for r in I]}:")
                         prof_I.anonymize().display()
                         print(prof_I.description())
                         vm.display(prof_I)
-                        print(f"\nThe profile after swapping {x} and {y} in the rankings {J}:")
+                        print(f"\nThe profile after swapping {x} and {y} in the rankings {[r.rmap for r in J]}:")
                         prof_J.anonymize().display()
                         print(prof_J.description())
                         vm.display(prof_J)
@@ -433,11 +433,11 @@ def find_all_preferential_equality_violations(prof, vm, verbose=False):
                         print("The original profile")
                         prof.anonymize().display()
                         print(prof.description())
-                        print(f"\nThe profile after swapping {x} and {y} in the rankings {I}:")
+                        print(f"\nThe profile after swapping {x} and {y} in the rankings {[r.rmap for r in I]}:")
                         prof_I.anonymize().display()
                         print(prof_I.description())
                         vm.display(prof_I)
-                        print(f"\nThe profile after swapping {x} and {y} in the rankings {J}:")
+                        print(f"\nThe profile after swapping {x} and {y} in the rankings {[r.rmap for r in J]}:")
                         prof_J.anonymize().display()
                         print(prof_J.description())
                         vm.display(prof_J)
@@ -459,11 +459,11 @@ def find_all_preferential_equality_violations(prof, vm, verbose=False):
                         print("The original profile")
                         prof.anonymize().display()
                         print(prof.description())
-                        print(f"\nThe profile after swapping {x} and {y} in the rankings {I}:")
+                        print(f"\nThe profile after swapping {x} and {y} in the rankings {[r.rmap for r in I]}:")
                         prof_I.anonymize().display()
                         print(prof_I.description())
                         vm.display(prof_I)
-                        print(f"\nThe profile after swapping {x} and {y} in the rankings {J}:")
+                        print(f"\nThe profile after swapping {x} and {y} in the rankings {[r.rmap for r in J]}:")
                         prof_J.anonymize().display()
                         print(prof_J.description())
                         vm.display(prof_J)
