### 2025-01-16 #7020 @britta-wstnr
Hi @epacuit
I went over the paper and everything looks fine to me! 🙂 Of course feel free to also check for typos and such again.
As you can see above, our bot identified two possible DOIs for papers you've cited. I believe the second one is wrong (different authors), but the first one seems to fit. Can you double check and potentially adapt? Thanks!

Otherwise, I agree with the bot that there are no DOIs for the other titles, but you've included URLs where possible.

Once the DOI thing is solved, we will enter our post-review checklist. 👇
