+++ paper/paper.bib
@@ -33,15 +33,22 @@
   doi        = {10.1007/978-3-319-91908-9_4}
 }
 
-@misc{Boehmer2024,
-  title      = {Guide to Numerical Experiments on Elections in Computational Social Choice},
-  author     = {Boehmer, N. and Faliszewski, P. and Janeczko, \L. and Kaczmarczyk, A. and Lisowski, G. and Pierczy\'{n}ski, G. and Rey, S. and Stolicki, D. and Szufa, S. and W\k{a}s, T.},
-  year       = {2024},
-  url        = {https://arxiv.org/abs/2402.11765},
-  eprint     = {2402.11765},
-  archivePrefix = {arXiv},
-  primaryClass  = {cs.GT}
+
+@inproceedings{Boehmer2024,
+  title     = {Guide to Numerical Experiments on Elections in Computational Social Choice},
+  author    = {Boehmer, N. and Faliszewski, P. and Janeczko, \L. and Kaczmarczyk, A. and Lisowski, G. and Pierczy\'{n}ski, G. and Rey, S. and Stolicki, D. and Szufa, S. and W\k{a}s, T.},
+    booktitle = {Proceedings of the Thirty-Third International Joint Conference on Artificial Intelligence, {IJCAI-24}},
+  publisher = {International Joint Conferences on Artificial Intelligence Organization},
+  editor    = {Kate Larson},
+  pages     = {7962--7970},
+  year      = {2024},
+  month     = {8},
+  note      = {Survey Track},
+  doi       = {10.24963/ijcai.2024/881},
+  url       = {https://doi.org/10.24963/ijcai.2024/881},
 }
+
+
 
 @incollection{Brams2002,
   title      = {Voting Procedures},
@@ -90,6 +97,8 @@
   address    = {Oxford},
   year       = {1984}
 }
+
+
 
 @incollection{Faliszewski2009,
   title      = {A richer understanding of the complexity of election systems},
