### 2025-01-16 71f96ba
update bib and add reference to JOSS paper
