+++ paper/paper.bib
@@ -174,16 +174,25 @@
   doi        = {10.1007/s00355-015-0909-0}
 }
 
-@inproceedings{HKP2024,
-  title      = {Learning to Manipulate under Limited Information},
-  author     = {Holliday, Wesley H. and Kristoffersen, Alexander and Pacuit, Eric},
-  booktitle  = {{1st Workshop on Social Choice and Learning Algorithms (SCaLA 2024) at the 23rd International Conference on Autonomous Agents and Multiagent Systems}},
-  editor     = {B. Armstrong and R. Fairstein and N. Mattei and Z. Terzopoulou},
-  address    = {Auckland, New Zealand},
-  month      = {May 6-7},
-  year       = {2024},
+@misc{HornischerTerzopoulou2024,
+      title={Learning How to Vote With Principles: Axiomatic Insights Into the Collective Decisions of Neural Networks}, 
+      author={Levin Hornischer and Zoi Terzopoulou},
+      year={2024},
+      eprint={2410.16170},
+      archivePrefix={arXiv},
+      primaryClass={cs.AI},
+      url={https://arxiv.org/abs/2410.16170}, 
+}
+
+@inproceedings{HKP2025,
+  title={Learning to Manipulate under Limited Information},
+  author={Wesley H. Holliday and Alexander Kristoffersen and Eric Pacuit},
+  year={Forthcoming},
+  note ={arXiv 2401.16412 [cs.AI]},
+  booktitle = {Proceedings of the 39th Annual {AAAI} Conference on Artificial Intelligence ({AAAI-25})},
   url        = {https://arxiv.org/abs/2401.16412}
 }
+
 
 @book{Kelly1988,
   title      = {Social Choice Theory: {A}n Introduction},
