+++ paper/paper.md
@@ -35,7 +35,7 @@ Preferential Voting Tools (`pref_voting`) is a Python package designed for resea
 
 # Statement of need
 
-Research in the burgeoning field of *computational social choice* (COMSOC) [@Brandt2016;@Geist2017;@Aziz2019] often applies computer-assisted techniques to the study of voting methods and other collective decision procedures. The aim of `pref_voting` is to contribute to a comprehensive set of tools for such research. Other packages in this area include `abcvoting` [@Lackner2023], which focuses on approval-based committee voting,  `preflibtools` [@Mattei2013], which contains tools for working with preference data from [PrefLib.org](https://PrefLib.org), and `prefsampling` [@Boehmer2024], which implements probability models for generating voter rankings. The `pref_voting` package provides functionality not available in these previous packages, while also interfacing with `preflibtools` and `prefsampling`. Like `pref_voting`, the `VoteKit` [@MGGG2024] and `VoteLib` [@Simbera2021] packages provide implementations of a number of voting methods; and like `prefsampling`, `VoteKit` provides tools for generating elections. However, neither package includes all the voting methods and functionality in `pref_voting`, as described below. The `pref_voting` package has already been used in COMSOC research [@HKP2024;@HornischerTerzopoulou2024] and in online COMSOC tools [@Peters2024]. The package can also be used by election administrators to determine election outcomes, as it is used in the [Stable Voting](https://stablevoting.org) website.
+Research in the burgeoning field of *computational social choice* (COMSOC) [@Brandt2016;@Geist2017;@Aziz2019] often applies computer-assisted techniques to the study of voting methods and other collective decision procedures. The aim of `pref_voting` is to contribute to a comprehensive set of tools for such research. Other packages in this area include `abcvoting` [@Lackner2023], which focuses on approval-based committee voting,  `preflibtools` [@Mattei2013], which contains tools for working with preference data from [PrefLib.org](https://PrefLib.org), and `prefsampling` [@Boehmer2024], which implements probability models for generating voter rankings. The `pref_voting` package provides functionality not available in these previous packages, while also interfacing with `preflibtools` and `prefsampling`. Like `pref_voting`, the `VoteKit` [@MGGG2024] and `VoteLib` [@Simbera2021] packages provide implementations of a number of voting methods; and like `prefsampling`, `VoteKit` provides tools for generating elections. However, neither package includes all the voting methods and functionality in `pref_voting`, as described below. The `pref_voting` package has already been used in COMSOC research [@HKP2025;@HornischerTerzopoulou2024] and in online COMSOC tools [@Peters2024]. The package can also be used by election administrators to determine election outcomes, as it is used in the [Stable Voting](https://stablevoting.org) website.
 
 # Functionality
 
