### 2025-01-16 #7020 @epacuit
I have double checked the authors and affiliations, and ORCIDs

The version number of the release is: 1.15.0

I have archived the release on Zenodo.   The DOI is: 10.5281/zenodo.14675584

I double checked that the title, the authors, and the License are all correct.
