### 2025-01-21 #7020 @britta-wstnr
Hi @epacuit, time for me to recommend acceptance for this paper! :sparkles: This means I hand the torch over to our EiC.

Big thanks to the reviewers @dmnapolitano and @pkrafft  for your work, I and JOSS appreciate it! :blue_heart:
Thanks everyone for the smooth work on this. :wave:
