### 2025-01-21 #7020 @dmnapolitano
> Hi [@epacuit](https://github.com/epacuit), time for me to recommend acceptance for this paper! ✨ This means I hand the torch over to our EiC.
>
> Big thanks to the reviewers [@dmnapolitano](https://github.com/dmnapolitano) and [@pkrafft](https://github.com/pkrafft) for your work, I and JOSS appreciate it! 💙 Thanks everyone for the smooth work on this. 👋

@britta-wstnr thanks so much to you and @pkrafft !  This was my first time reviewing for JOSS and I learned a lot about the process from both of you 😄 🎉 👏🏻
