### 2025-01-22 #7020 @epacuit
Sorry, but there is one thing that we just noticed, there is a bib entry that has an incorrect date:

Holliday, W. H., Kristoffersen, A., & Pacuit, E. (ForthcomingForthcoming). Learning to
manipulate under limited information. Proceedings of the 39th Annual AAAI Conference
on Artificial Intelligence (AAAI-25). https://arxiv.org/abs/2401.16412

Do you know how to make sure it says "Forthcoming" rather than "ForthcomingForthcoming".

The entry in the bib file is correct:

@inproceedings{HKP2025,
  title={Learning to Manipulate under Limited Information},
  author={Wesley H. Holliday and Alexander Kristoffersen and Eric Pacuit},
  year={Forthcoming},
  note ={arXiv 2401.16412 [cs.AI]},
  booktitle = {Proceedings of the 39th Annual {AAAI} Conference on Artificial Intelligence ({AAAI-25})},
  url        = {https://arxiv.org/abs/2401.16412}
}

Usually this works fine, but is there a requirement that "year" field is a number?
