+++ paper/paper.bib
@@ -187,7 +187,7 @@
 @inproceedings{HKP2025,
   title={Learning to Manipulate under Limited Information},
   author={Wesley H. Holliday and Alexander Kristoffersen and Eric Pacuit},
-  year={test},
+  year={2025},
   note ={arXiv 2401.16412 [cs.AI]},
   booktitle = {Proceedings of the 39th Annual {AAAI} Conference on Artificial Intelligence ({AAAI-25})},
   url        = {https://arxiv.org/abs/2401.16412}
