### 2025-01-22 18807e1
update bib so all year fields are integers
