### 2025-01-22 #7020 @epacuit
I think that there is a bug when processing the bib file.   If the year field is a string, then it will write the field twice.  To test, this I set the year field to the string "2025" (i.e., I used year={"2025"}), and this is what is produced:

Holliday, W. H., Kristoffersen, A., & Pacuit, E. (”2025””2025”). Learning to manipulate
143 under limited information. Proceedings of the 39th Annual AAAI Conference on Artificial
144 Intelligence (AAAI-25). https://arxiv.org/abs/2401.16412

We can fix this by setting the year to 2025 (the paper is coming out at the next AAAI meeting in a little over a month...).

I've done that now, so can we use the most recent version in our repo.  Thanks!!
