### 2025-01-23 #7020 @samhforbes
Whew, I can't see immediately why it does that. If it's an ongoing issue we can tag in the dev team, but sounds like it's not an issue for now!
