### 2025-01-25 dde8ec1
Updated label options when viewing spatial profile
Separated show_labels into show_cand_labels and show_voter_labels
