### 2025-01-25 #7020 @epacuit
I just wanted to check if there is anything else I need to do to ensure that the issue with the bib file is corrected in the final version of the paper?   The latest version in our repo removed the "Forthcoming" and replaced with with an integer so to bypass the bug with processing the bib file.   Thanks!
