### 2025-01-28 #7020 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
